% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{a_rungekutta}
\alias{a_rungekutta}
\title{Runge Kutta for the calculation of the a vectors in a EM step}
\usage{
a_rungekutta(avector, dt, h, T)
}
\arguments{
\item{avector}{the a vector}

\item{dt}{increment}

\item{h}{step-length}

\item{T}{sub-intensity}
}
\description{
Can be used for the loglikelihood
}
