% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{a_rungekutta_new}
\alias{a_rungekutta_new}
\title{Runge Kutta for the calculation of the a vectors in a EM step}
\usage{
a_rungekutta_new(avector, dt, h, S)
}
\arguments{
\item{avector}{The a vector.}

\item{dt}{Increment.}

\item{h}{Step-length.}

\item{S}{Sub-intensity.}
}
\description{
Runge Kutta for the calculation of the a vectors in a EM step
}
