% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2c.iph_aft.R
\name{aft,ph-method}
\alias{aft,ph-method}
\title{AFT regression Method for ph Class}
\usage{
\S4method{aft}{ph}(
  x,
  y,
  weight = numeric(0),
  rcen = numeric(0),
  rcenweight = numeric(0),
  X = numeric(0),
  B0 = numeric(0),
  stepsEM = 1000
)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{vector or data.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
AFT regression Method for ph Class
}
\examples{

}
