% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{aft}
\alias{aft}
\title{New Generic for AFT regression with Matrix Distributions}
\usage{
aft(x, y, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{y}{a vector of data.}

\item{...}{}
}
\value{
An object of the fitted model class.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
