% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bimParetotail}
\alias{bimParetotail}
\title{Bivariate matrix Weibull joint tail}
\usage{
bimParetotail(x, alpha, T11, T12, T22, beta)
}
\description{
Bivariate matrix Weibull joint tail
}
\examples{
alpha <- c(0.15, 0.85)
T11 <- matrix(c(c(-2,9),c(0,-11)), nrow = 2, ncol = 2)
T12 <- matrix(c(c(2,0),c(0,2)), nrow = 2, ncol = 2)
T22 <- matrix(c(c(-1,0),c(0.5,-5)), nrow = 2, ncol = 2)
beta <- c(2, 4)
x1 <- matrix(c(0.5,1), ncol=2) 
x2 <- matrix(c(c(0.5,1), c(2, 1.5)), ncol=2) 
bimPartail(x1, alpha, T11, T12, T22, beta) 
bimPartail(x2, alpha, T11, T12, T22, beta) 
}
