% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\docType{class}
\name{bivdph-class}
\alias{bivdph-class}
\title{Bivariate discrete phase-type distributions}
\value{
Class object.
}
\description{
Class of objects for bivariate discrete phase-type distributions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the discrete phase-type distribution.}

\item{\code{pars}}{A list comprising of the parameters.}

\item{\code{fit}}{A list containing estimation information.}
}}

