% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4c.biviph.R
\docType{class}
\name{biviph-class}
\alias{biviph-class}
\title{Bivariate inhomogeneous phase-type distributions}
\value{
Class object.
}
\description{
Class of objects for bivariate inhomogeneous phase-type distributions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the phase type distribution.}

\item{\code{gfun}}{A list comprising of the parameters.}
}}

