% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bivph_density}
\alias{bivph_density}
\title{Bivariate phase-type joint density of the feed forward type}
\usage{
bivph_density(x, alpha, S11, S12, S22)
}
\arguments{
\item{x}{Matrix of values.}

\item{alpha}{Vector of initial probabilities.}

\item{S11}{Sub-intensity matrix.}

\item{S12}{Matrix.}

\item{S22}{Sub-intensity matrix.}
}
\value{
Joint density at \code{x}.
}
\description{
Bivariate phase-type joint density of the feed forward type
}
