% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bivph_laplace}
\alias{bivph_laplace}
\title{Bivariate phase-type joint Laplace}
\usage{
bivph_laplace(r, alpha, S11, S12, S22)
}
\arguments{
\item{r}{Matrix of values.}

\item{alpha}{Vector of initial probabilities.}

\item{S11}{Sub-intensity matrix.}

\item{S12}{Matrix.}

\item{S22}{Sub-intensity matrix.}
}
\value{
Joint laplace at \code{r}.
}
\description{
Bivariate phase-type joint Laplace
}
