% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bph.R
\docType{class}
\name{bph-class}
\alias{bph-class}
\title{Bivariate Phase Type distributions}
\value{

}
\description{
Class of objects for bivariate phase type distributions
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the phase type distribution}

\item{\code{pars}}{a list comprising of the parameters.}
}}

