% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bph.R
\name{bph}
\alias{bph}
\title{Constructor Function for bivariate phase type distributions}
\usage{
bph(
  alpha = NULL,
  T11 = NULL,
  T12 = NULL,
  T22 = NULL,
  rand_str = FALSE,
  dimensions = c(3, 3)
)
}
\arguments{
\item{alpha}{a probability vector.}

\item{dimensions}{the dimensions of the ph structures (if provided)}

\item{S}{a sub-intensity matrix.}

\item{structure}{a valid ph structure}
}
\value{
An object of class \linkS4class{bph}.
}
\description{
Constructor Function for bivariate phase type distributions
}
