% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bph.R
\name{cdf,bph-method}
\alias{cdf,bph-method}
\title{Distribution Method for bivariate phase type distributions}
\usage{
\S4method{cdf}{bph}(x, q = seq(0, quan(x, 0.95)$quantile, length.out = 10), lower.tail = TRUE)
}
\arguments{
\item{x}{an object of class \linkS4class{bph}.}

\item{y}{locations}
}
\value{
CDF evaluated at locations
}
\description{
Distribution Method for bivariate phase type distributions
}
\examples{

}
