% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{cdf,dph-method}
\alias{cdf,dph-method}
\title{Distribution method for discrete phase-type distributions}
\usage{
\S4method{cdf}{dph}(x, q, lower.tail = TRUE)
}
\arguments{
\item{x}{An object of class \linkS4class{dph}.}

\item{q}{A vector of locations.}

\item{lower.tail}{Logical parameter specifying whether lower tail (CDF) or
upper tail is computed.}
}
\value{
A vector containing the CDF evaluations at the given locations.
}
\description{
Distribution method for discrete phase-type distributions
}
\examples{
obj <- dph(structure = "general")
cdf(obj, c(1, 2, 3))
}
