% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.ibph.R
\name{cdf,ibph-method}
\alias{cdf,ibph-method}
\title{Distribution Method for inhomogeneous bivariate phase type distributions}
\usage{
\S4method{cdf}{ibph}(x, q = seq(0, quan(x, 0.95)$quantile, length.out = 10), lower.tail = TRUE)
}
\arguments{
\item{x}{an object of class \linkS4class{ibph}.}

\item{y}{locations}
}
\value{
CDF evaluated at locations
}
\description{
Distribution Method for inhomogeneous bivariate phase type distributions
}
\examples{

}
