% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{cdf,iph-method}
\alias{cdf,iph-method}
\title{Distribution Method for inhomogeneous phase type distributions}
\usage{
\S4method{cdf}{iph}(x, q = seq(0, quan(x, 0.95)$quantile, length.out = 10), lower.tail = TRUE)
}
\arguments{
\item{x}{an object of class \linkS4class{iph}.}

\item{q}{locations}

\item{lower.tail}{cdf(TRUE) or tail(FALSE)}
}
\value{
CDF evaluated at locations
}
\description{
Distribution Method for inhomogeneous phase type distributions
}
