% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{cdf,mph-method}
\alias{cdf,mph-method}
\title{Distribution Method for multivariate phase type distributions}
\usage{
\S4method{cdf}{mph}(x, q, lower.tail = TRUE)
}
\arguments{
\item{x}{an object of class \linkS4class{mph}.}

\item{q}{a vector of locations.}

\item{lower.tail}{logical parameter specifying whether lower tail (cdf) or upper tail is computed.}
}
\value{
A list containing the locations and corresponding CDF evaluations.
}
\description{
Distribution Method for multivariate phase type distributions
}
