% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{cdf,ph-method}
\alias{cdf,ph-method}
\title{Distribution Method for phase type distributions}
\usage{
\S4method{cdf}{ph}(x, q = seq(0, quan(x, 0.95)$quantile, length.out = 10), lower.tail = TRUE)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{q}{locations}
}
\value{
CDF evaluated at locations
}
\description{
Distribution Method for phase type distributions
}
\examples{

}
