% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{cdf}
\alias{cdf}
\title{New Generic for the Distribution of Matrix Distributions}
\usage{
cdf(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
CDF from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
