% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\name{coef,bivdph-method}
\alias{coef,bivdph-method}
\title{Coef method for bivdph class}
\usage{
\S4method{coef}{bivdph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{bivdph}.}
}
\value{
Parameters of bivariate discrete phase-type model.
}
\description{
Coef method for bivdph class
}
\examples{
obj <- bivdph(dimensions = c(3, 3))
coef(obj)
}
