% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.bivph.R
\name{coef,bivph-method}
\alias{coef,bivph-method}
\title{Coef method for bivph class}
\usage{
\S4method{coef}{bivph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{bivph}.}
}
\value{
Parameters of bivariate phase-type model.
}
\description{
Coef method for bivph class
}
\examples{
obj <- bivph(dimensions = c(3, 3))
coef(obj)
}
