% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{coef,dph-method}
\alias{coef,dph-method}
\title{Coef method for dph Class}
\usage{
\S4method{coef}{dph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{dph}.}
}
\value{
Parameters of dph model.
}
\description{
Coef method for dph Class
}
\examples{
obj <- dph(structure = "general", dim = 3)
coef(obj)
}
