% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{coef,mdph-method}
\alias{coef,mdph-method}
\title{Coef method for mdph class}
\usage{
\S4method{coef}{mdph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{mdph}.}
}
\value{
Parameters of multivariate discrete phase-type model.
}
\description{
Coef method for mdph class
}
\examples{
obj <- mdph(structure = c("general", "general"))
coef(obj)
}
