% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{coef,ph-method}
\alias{coef,ph-method}
\title{Coef Method for ph Class}
\usage{
\S4method{coef}{ph}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{ph}.}
}
\value{
parameters of ph model
}
\description{
Coef Method for ph Class
}
