% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.sph.R
\name{coef,sph-method}
\alias{coef,sph-method}
\title{Coef Method for sph Class}
\usage{
\S4method{coef}{sph}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{sph}.}
}
\value{
parameters of sph model
}
\description{
Coef Method for sph Class
}
