% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\name{cor,bivdph-method}
\alias{cor,bivdph-method}
\title{Cor method for bivdph class}
\usage{
\S4method{cor}{bivdph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{bivdph}.}
}
\value{
The correlation matrix of the bivariate discrete phase-type distribution.
}
\description{
Cor method for bivdph class
}
\examples{
obj <- bivdph(dimensions = c(3, 3))
cor(obj)
}
