% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{cor,mdph-method}
\alias{cor,mdph-method}
\title{Cor method for multivariate discrete phase-type distributions}
\usage{
\S4method{cor}{mdph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{mdph}.}
}
\value{
The correlation matrix of the multivariate discrete phase-type distribution.
}
\description{
Cor method for multivariate discrete phase-type distributions
}
\examples{
obj <- mdph(structure = c("general", "general"))
cor(obj)
}
