% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{cor,mph-method}
\alias{cor,mph-method}
\title{Cor method for multivariate phase-type distributions}
\usage{
\S4method{cor}{mph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}
}
\value{
The correlation matrix of the multivariate phase-type distribution.
}
\description{
Cor method for multivariate phase-type distributions
}
\examples{
obj <- mph(structure = c("general", "general"))
cor(obj)
}
