% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{cor,MPHstar-method}
\alias{cor,MPHstar-method}
\title{Cor method for MPHstar class}
\usage{
\S4method{cor}{MPHstar}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{MPHstar}.}
}
\value{
The correlation matrix of the MPHstar distribution.
}
\description{
Cor method for MPHstar class
}
\examples{
obj <- MPHstar(structure = "general")
cor(obj)
}
