% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{csph_densityrk}
\alias{csph_densityrk}
\title{Bivariate phase-type joint density of the common shock type}
\usage{
csph_densityrk(x, alpha, S, P, Q1, Q2, h)
}
\arguments{
\item{x}{Matrix of values.}

\item{alpha}{Vector of initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{P}{Matrix.}

\item{Q1}{Sub-intensity matrix.}

\item{Q2}{Sub-intensity matrix.}

\item{h}{step size.}
}
\value{
Joint density at \code{x}.
}
\description{
Bivariate phase-type joint density of the common shock type
}
