% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumulate_matrix}
\alias{cumulate_matrix}
\title{Cumulate matrix}
\usage{
cumulate_matrix(A)
}
\arguments{
\item{A}{A matrix.}
}
\value{
The cumulated matrix.
}
\description{
Creates a new matrix with entries the cumulated rows of \code{A}.
}
