% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumulate_vector}
\alias{cumulate_vector}
\title{Cumulate vector}
\usage{
cumulate_vector(A)
}
\arguments{
\item{A}{A vector.}
}
\value{
The cumulated vector.
}
\description{
Creates a new vector with entries the cumulated entries of \code{A}.
}
