% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumulateMatrix}
\alias{cumulateMatrix}
\title{Cumulate matrix}
\usage{
cumulateMatrix(A)
}
\arguments{
\item{A}{A matrix}
}
\value{
The cumulated matrix
}
\description{
Creates a new matrix with entries the cumulated rows of \code{A}
}
