% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumulateVector}
\alias{cumulateVector}
\title{Cumulate vector}
\usage{
cumulateVector(A)
}
\arguments{
\item{A}{A vector}
}
\value{
The cumulated vector
}
\description{
Creates a new vector with entries the cumulated entries of \code{A}
}
