% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{d,ph-method}
\alias{d,ph-method}
\title{Density Method for phase type distributions}
\usage{
\S4method{d}{ph}(x, y = seq(0, 5, length.out = 100))
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{locations}
}
\value{
Density evaluated at locations
}
\description{
Density Method for phase type distributions
}
\examples{

}
