% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{d}
\alias{d}
\title{New Generic for the Density of Matrix Distributions}
\usage{
d(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{}
}
\value{
Density from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
