% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R, R/2b.iph_regression.R,
%   R/2c.iph_inhomogenity_regression.R
\name{data_aggregation}
\alias{data_aggregation}
\title{Aggregate data to improve computation speed}
\usage{
data_aggregation(y, w)

data_aggregation(y, w)

data_aggregation(y, w)
}
\arguments{
\item{y}{Observations. Either a vector or a matrix.}

\item{w}{Respective weights of observations}
}
\value{
Returns a named list with unique observations and associated weights. If y is a vector then the unique observations are given, otherwise the unique rows are returned.

Returns a named list with unique observations and associated weights. If y is a vector then the unique observations are given, otherwise the unique rows are returned.

Returns a named list with unique observations and associated weights. If y is a vector then the unique observations are given, otherwise the unique rows are returned.
}
\description{
Aggregate data to improve computation speed

Aggregate data to improve computation speed

Aggregate data to improve computation speed
}
