% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{default_step_length}
\alias{default_step_length}
\title{Default size of the steps in the RK}
\usage{
default_step_length(S)
}
\arguments{
\item{S}{sub-intensity matrix}
}
\value{
The step length for \code{S}
}
\description{
Computes the default step length for a matrix \code{S} to be employed in the RK method
}
