% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\name{dens,bivdph-method}
\alias{dens,bivdph-method}
\title{Density method for bivariate discrete phase-type distributions}
\usage{
\S4method{dens}{bivdph}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{bivdph}.}

\item{y}{A matrix of locations.}
}
\value{
A vector containing the joint density evaluations at the given locations.
}
\description{
Density method for bivariate discrete phase-type distributions
}
\examples{
obj <- bivdph(dimensions = c(3, 3))
dens(obj, matrix(c(1, 2), ncol = 2))
}
