% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.bivph.R
\name{dens,bivph-method}
\alias{dens,bivph-method}
\title{Density method for bivariate phase-type distributions}
\usage{
\S4method{dens}{bivph}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{bivph}.}

\item{y}{A matrix of locations.}
}
\value{
A vector containing the joint density evaluations at the given locations.
}
\description{
Density method for bivariate phase-type distributions
}
\examples{
obj <- bivph(dimensions = c(3, 3))
dens(obj, matrix(c(0.5, 1), ncol = 2))
}
