% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bph.R
\name{dens,bph-method}
\alias{dens,bph-method}
\title{Density Method for bivariate phase type distributions}
\usage{
\S4method{dens}{bph}(x, y = matrix(c(1, 2, 1, 2), 2, 2))
}
\arguments{
\item{x}{an object of class \linkS4class{bph}.}

\item{y}{locations}
}
\value{
Density evaluated at locations
}
\description{
Density Method for bivariate phase type distributions
}
\examples{

}
