% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{dens,dph-method}
\alias{dens,dph-method}
\title{Density method for discrete phase-type distributions}
\usage{
\S4method{dens}{dph}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{dph}.}

\item{y}{A vector of locations.}
}
\value{
A vector containing the density evaluations at the given locations.
}
\description{
Density method for discrete phase-type distributions
}
\examples{
obj <- dph(structure = "general")
dens(obj, c(1, 2, 3))
}
