% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.ibph.R
\name{dens,ibph-method}
\alias{dens,ibph-method}
\title{Density Method for inhomogeneous bivariate phase type distributions}
\usage{
\S4method{dens}{ibph}(x, y = matrix(c(1, 2, 1, 2), 2, 2))
}
\arguments{
\item{x}{an object of class \linkS4class{ibph}.}

\item{y}{locations}
}
\value{
Density evaluated at locations
}
\description{
Density Method for inhomogeneous bivariate phase type distributions
}
\examples{

}
