% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{dens,mdph-method}
\alias{dens,mdph-method}
\title{Density method for multivariate discrete phase-type distributions}
\usage{
\S4method{dens}{mdph}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{mdph}.}

\item{y}{A matrix of locations.}
}
\value{
A vector containing the joint density evaluations at the given locations.
}
\description{
Density method for multivariate discrete phase-type distributions
}
\examples{
obj <- mdph(structure = c("general", "general"))
dens(obj, matrix(c(1, 1), ncol = 2))
}
