% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{dens,ph-method}
\alias{dens,ph-method}
\title{Density Method for phase type distributions}
\usage{
\S4method{dens}{ph}(x, y = seq(0, quan(x, 0.95)$quantile, length.out = 10))
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{locations}
}
\value{
Density evaluated at locations
}
\description{
Density Method for phase type distributions
}
\examples{
obj <- ph(structure = "General")
dens(obj, c(1, 2, 3))
}
