% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derivativeMatrixWeibull}
\alias{derivativeMatrixWeibull}
\title{Derivative of matrix Weibull}
\usage{
derivativeMatrixWeibull(h, obs, weight, rcens, rcweight, pi, T, beta)
}
\arguments{
\item{h}{step-lenght}

\item{obs}{the observations}

\item{weight}{weight of the observations}

\item{rcens}{censored observations}

\item{rcweight}{weight of the censored observations}

\item{pi}{initial probabilities}

\item{T}{sub-intensity}

\item{beta}{parametor of transformation}
}
\description{
Can be used to increase performance
}
