% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixnorm.R
\name{dmatrixnorm.test}
\alias{dmatrixnorm.test}
\title{dmatrixnorm.test}
\usage{
dmatrixnorm.test(x, mean = array(0L, dim(x)), L = diag(dim(x)[1]),
  R = diag(dim(x)[2]), U = L \%*\% t(L), V = t(R) \%*\% R, log = FALSE)
}
\arguments{
\item{x}{\eqn{p X q} input matrix}

\item{mean}{\eqn{p X q} matrix of means. By default, a matrix of \eqn{0}s with size taken from \code{x}}

\item{L}{\eqn{p X p} matrix specifying relations among the rows. By default, an identity matrix.}

\item{R}{\eqn{q X q} matrix specifying relations among the columns. By default, an identity matrix.}

\item{U}{\eqn{LL^T} - \eqn{p X p} positive definite variance-covariance matrix for rows, computed from \eqn{L} if not specified.}

\item{V}{\eqn{R^T R} - \eqn{q X q} positive definite variance-covariance matrix for columns, computed from \eqn{R} if not specified.}

\item{log}{Whether to return the density on the log scale.}
}
\value{
Returns the density at the provided observation. This is an alternative method of computing which works by flattening out into a vector instead of a matrix.
}
\description{
Equivalent to dmatrixnorm except it works by unrolling to a vector.
}
\examples{
set.seed(20180202)
A = rmatrixnorm(n=1,mean=matrix(c(100,0,-100,0,25,-1000),nrow=2),
    L=matrix(c(2,1,0,.1),nrow=2))
\dontrun{dmatrixnorm.test (A,mean=matrix(c(100,0,-100,0,25,-1000),nrow=2),
    L=matrix(c(2,1,0,.1),nrow=2),log=TRUE )}
}
\keyword{#'}
\keyword{internal}
