% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\docType{class}
\name{dph-class}
\alias{dph-class}
\title{Discrete Phase Type distributions}
\value{
Class object.
}
\description{
Class of objects for discrete phase-type distributions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the discrete phase-type distribution.}

\item{\code{pars}}{A list comprising of the parameters.}

\item{\code{fit}}{A list containing estimation information.}
}}

