% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dph_pgf}
\alias{dph_pgf}
\title{Pgf of a discrete phase-type distribution}
\usage{
dph_pgf(z, alpha, S)
}
\arguments{
\item{z}{Vector of real values.}

\item{alpha}{Vector of initial probabilities.}

\item{S}{Sub-transition matrix.}
}
\value{
Laplace transform at \code{r}.
}
\description{
Computes the pgf at \code{z} of a discrete phase-type distribution with
 parameters \code{alpha} and \code{S}.
}
