% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dphdensity}
\alias{dphdensity}
\title{Discrete phase-type density}
\usage{
dphdensity(x, alpha, S)
}
\arguments{
\item{x}{Non-negative value.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-transition matrix.}
}
\value{
The density at \code{x}.
}
\description{
Computes the density of discrete phase-type distribution with parameters
\code{alpha} and \code{S} at \code{x}.
}
