% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EM_step_mPH_rc}
\alias{EM_step_mPH_rc}
\title{EM step for the mPH class with right-censoring, for different marginal sub-intensity matrices}
\usage{
EM_step_mPH_rc(alpha, S_list, y, delta, h)
}
\arguments{
\item{alpha}{Common initial distribution vector}

\item{S_list}{List of marginal sub-intensity matrices}

\item{y}{Matrix of marginal observations}

\item{h}{Tolerance of uniformization}
}
\description{
EM step for the mPH class with right-censoring, for different marginal sub-intensity matrices
}
