% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{embedded_mc}
\alias{embedded_mc}
\title{Embedded Markov chain of a sub-intensity matrix}
\usage{
embedded_mc(S)
}
\arguments{
\item{S}{A sub-intensity matrix.}
}
\value{
The embedded Markov chain.
}
\description{
Returns the transition probabilities of the embedded Markov chain determined
the sub-intensity matrix.
}
