% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep}
\alias{EMstep}
\title{EM using Matlab algorithm for matrix exponential}
\usage{
EMstep(pi, T, obs, weight, rcens, rcweight)
}
\description{
This one is slower but dont requires to order the sample
}
