% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_bivdph}
\alias{EMstep_bivdph}
\title{EM for discrete bivariate phase-type}
\usage{
EMstep_bivdph(alpha, S11, S12, S22, obs, weight)
}
\arguments{
\item{alpha}{Initial probabilities.}

\item{S11}{Sub-transition matrix.}

\item{S12}{Matrix.}

\item{S22}{Sub-transition matrix.}

\item{obs}{The observations.}

\item{weight}{The weights for the observations.}
}
\description{
EM for discrete bivariate phase-type
}
