% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_bivph}
\alias{EMstep_bivph}
\title{EM for bivariate phase-type distributions using Pade for matrix exponential}
\usage{
EMstep_bivph(alpha, S11, S12, S22, obs, weight)
}
\arguments{
\item{alpha}{Initial probabilities.}

\item{S11}{Sub-intensity.}

\item{S12}{A matrix.}

\item{S22}{Sub-intensity.}

\item{obs}{The observations.}

\item{weight}{The weights for the observations.}
}
\value{
Fitted alpha, S11, S12 and S22 after one iteration.
}
\description{
EM for bivariate phase-type distributions using Pade for matrix exponential
}
