% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_mdph_MoE}
\alias{EMstep_mdph_MoE}
\title{EM for multivariate discrete phase-type MoE}
\usage{
EMstep_mdph_MoE(alpha, S_list, obs, weight)
}
\arguments{
\item{alpha}{Initial probabilities.}

\item{S_list}{List of marginal sub-transition matrices.}

\item{obs}{The observations.}

\item{weight}{The weights for the observations.}
}
\description{
EM for multivariate discrete phase-type MoE
}
