% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_MoE_PADE}
\alias{EMstep_MoE_PADE}
\title{EM for PH-MoE}
\usage{
EMstep_MoE_PADE(alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{obs}{The observations.}

\item{weight}{The weights for the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{The weights for the censored observations.}
}
\description{
No recycling of information
}
