% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.sph.R
\name{eval,sph-method}
\alias{eval,sph-method}
\title{Evaluation Method for sph Class}
\usage{
\S4method{eval}{sph}(x, subject)
}
\arguments{
\item{object}{an object of class \linkS4class{sph}.}
}
\value{
a ph model
}
\description{
Evaluation Method for sph Class
}
