% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{eval}
\alias{eval}
\title{New Generic for Evaluating Survival Matrix Distributions}
\usage{
eval(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{}

\item{y}{a vector of data.}
}
\description{
Methods are available for objects of class \linkS4class{sph}
}
