% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.sph.R
\name{evaluate,sph-method}
\alias{evaluate,sph-method}
\title{Evaluation Method for sph Class}
\usage{
\S4method{evaluate}{sph}(x, subject)
}
\arguments{
\item{x}{an object of class \linkS4class{sph}.}

\item{subject}{covariates of a single subject.}
}
\value{
a \linkS4class{ph} model
}
\description{
Evaluation Method for sph Class
}
