% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{evaluate}
\alias{evaluate}
\title{New Generic for Evaluating Survival Matrix Distributions}
\usage{
evaluate(x, subject, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{subject}{A vector of data.}

\item{...}{Further parameters to be passed on.}
}
\description{
Methods are available for objects of class \linkS4class{sph}.
}
