% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{find_weight}
\alias{find_weight}
\title{Find weight of observations}
\usage{
find_weight(x)
}
\arguments{
\item{x}{A vector of observations from which we want to know their weights.}
}
\value{
A matrix with unique observations as first column and associated weights for second column.
}
\description{
Find weight of observations
}
