% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findN}
\alias{findN}
\title{Find n such that P(N > n) = h with N Poisson distributed}
\usage{
findN(h, lambda)
}
\arguments{
\item{h}{Probability.}

\item{lambda}{Mean of Poisson random variable.}
}
\value{
Integer satisfying condition.
}
\description{
Find n such that P(N > n) = h with N Poisson distributed
}
