% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{Fisher}
\alias{Fisher}
\title{New generic for obtaining the Fisher information of survival matrix distributions}
\usage{
Fisher(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\description{
Methods are available for objects of class \linkS4class{sph}.
}
