% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bph.R
\name{fit,bph-method}
\alias{fit,bph-method}
\title{Fit Method for bph Class}
\usage{
\S4method{fit}{bph}(x, y, weight = numeric(0), stepsEM = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{bph}.}

\item{y}{vector or data.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Fit Method for bph Class
}
\examples{

}
