% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{fit,mdph-method}
\alias{fit,mdph-method}
\title{Fit method for mdph Class}
\usage{
\S4method{fit}{mdph}(x, y, weight = numeric(0), stepsEM = 1000, every = 10)
}
\arguments{
\item{x}{An object of class \linkS4class{mdph}.}

\item{y}{A matrix with the data.}

\item{weight}{Vector of weights.}

\item{stepsEM}{Number of EM steps to be performed.}

\item{every}{Number of iterations between likelihood display updates.}
}
\value{
An object of class \linkS4class{mdph}.
}
\description{
Fit method for mdph Class
}
\examples{
obj <- mdph(structure = c("general", "general"))
data <- sim(obj, n = 100)
fit(obj, data, stepsEM = 100, every = 50)
}
