% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{fit}
\alias{fit}
\title{New Generic for Estimating Matrix Distributions}
\usage{
fit(x, y, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{y}{a vector of data.}

\item{...}{further parameters to be passed on}
}
\value{
An object of the fitted model class.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
