% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{haz}
\alias{haz}
\title{New Generic for the Hazard rate of Matrix Distributions}
\usage{
haz(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{further parameters to be passed on}
}
\value{
Hazard rate from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
