% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.ibph.R
\docType{class}
\name{ibph-class}
\alias{ibph-class}
\title{Inhomogeneous Bivariate Phase Type distributions}
\value{

}
\description{
Class of objects for inhomogeneous bivariate phase type distributions
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the phase type distribution}

\item{\code{pars}}{a list comprising of the parameters.}
}}

