% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.ibph.R
\name{ibph}
\alias{ibph}
\title{Constructor Function for inhomogeneous bivariate phase type distributions}
\usage{
ibph(
  bph = NULL,
  gfun = NULL,
  gfun_pars = NULL,
  alpha = NULL,
  T11 = NULL,
  T12 = NULL,
  T22 = NULL,
  structure = NULL,
  dimensions = c(3, 3)
)
}
\arguments{
\item{alpha}{a probability vector.}

\item{structure}{a valid ph structure}

\item{S}{a sub-intensity matrix.}

\item{dimension}{the dimension of the ph structure (if provided)}

\item{fun}{inhomogeneity transform}

\item{funpars}{the parameters of the inhomogeneity function}
}
\value{
An object of class \linkS4class{ibph}.
}
\description{
Constructor Function for inhomogeneous bivariate phase type distributions
}
