% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inf_norm}
\alias{inf_norm}
\title{L inf norm of a matrix}
\usage{
inf_norm(A)
}
\arguments{
\item{A}{A matrix.}
}
\value{
The L inf norm.
}
\description{
Computes the L inf norm of a matrix \code{A}, which is defined as:
L_inf(A) =  max(1 <= i <= M) sum(1 <= j <= N) abs(A(i,j)).
}
