% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{inhomogeneity_reg}
\alias{inhomogeneity_reg}
\title{New generic for regression with matrix distributions, when the inhomogeneity parameter is also regressed}
\usage{
inhomogeneity_reg(x, y, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{y}{A vector of data.}

\item{...}{Further parameters to be passed on.}
}
\value{
An object of the fitted model class.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
