% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{initialState}
\alias{initialState}
\title{Initial state of Markov jump process}
\usage{
initialState(cumulatedPi, u)
}
\arguments{
\item{cumulatedPi}{A vector}

\item{u}{A random value in (0,1)}
}
\value{
The initial state of the Markov jump process
}
\description{
Given the accumulated values of the initial probabilities \code{Pi} and a uniform value \code{u}, it returns the initial state of a Markov jump process
}
