% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\docType{class}
\name{iph-class}
\alias{iph-class}
\title{Inhomogeneous Phase Type distributions}
<<<<<<< HEAD
=======
\value{

}
>>>>>>> bfd4a4437aecc2d9b403fc560a9345bfda8bd07c
\description{
Class of objects for inhomogeneous phase type distributions
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the phase type distribution}

\item{\code{pars}}{a list comprising of the parameters.}
}}

