% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{iph}
\alias{iph}
\title{Constructor Function for inhomogeneous phase type distributions}
\usage{
<<<<<<< HEAD
iph(ph = NULL, gfun = NULL, gfun_pars = NULL, alpha = NULL,
  S = NULL, structure = NULL, dimension = 3)
=======
iph(
  ph = NULL,
  gfun = NULL,
  gfun_pars = NULL,
  alpha = NULL,
  S = NULL,
  structure = NULL,
  dimension = 3
)
>>>>>>> bfd4a4437aecc2d9b403fc560a9345bfda8bd07c
}
\arguments{
\item{alpha}{a probability vector.}

\item{S}{a sub-intensity matrix.}

\item{structure}{a valid ph structure}

\item{dimension}{the dimension of the ph structure (if provided)}

\item{fun}{inhomogeneity transform}

\item{funpars}{the parameters of the inhomogeneity function}
}
\value{
An object of class \linkS4class{iph}.
}
\description{
Constructor Function for inhomogeneous phase type distributions
}
