% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iphcdf}
\alias{iphcdf}
\title{IPH cdf (tail)}
\usage{
iphcdf(x, pi, T, g, g_inv, beta, lower_tail = TRUE)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{g}{Tranformation}

\item{g_inv}{Inverse of the transformation}

\item{beta}{parameter of the transformation}

\item{lambda}{Derivative of the inverse}
}
\value{
The cdf (tail) at \code{x}
}
\description{
Computes the cdf(tail) of an IPH distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
}
\examples{
g <- function(x, beta) { x^(1/beta) }
g_inv <- function(x, beta) { x^beta}
alpha <- c(0.5, 0.3, 0.2)
T <- matrix(c(c(-1,0,0),c(1,-2,0),c(0,1,-5)), nrow = 3, ncol = 3)
beta <- 0.5
iphcdf(0.5, alpha, T, g, g_inv, beta)
iphcdf(0.5, alpha, T, g, g_inv, beta, FALSE) 
}
