% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{iphdensity}
\alias{iphdensity}
\title{IPH density - Slower}
\usage{
iphdensity(x, pi, T, g, g_inv, lambda, beta)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{g}{Tranformation}

\item{g_inv}{Inverse of the transformation}

\item{lambda}{Derivative of the inverse}

\item{beta}{parameter of the transformation}
}
\value{
The density at \code{x}
}
\description{
Computes the density of an IPH distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
}
\examples{
g <- function(x, beta) { x^(1/beta) }
g_inv <- function(x, beta) { x^beta}
lambda <- function(x, beta) {beta * x^(beta - 1)}
alpha <- c(0.5, 0.3, 0.2)
T <- matrix(c(c(-1,0,0),c(1,-2,0),c(0,1,-5)), nrow = 3, ncol = 3)
beta <- 0.5
iphdensity(0.5, alpha, T, g, g_inv, lambda, beta) 
}
