% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{laplace,mph-method}
\alias{laplace,mph-method}
\title{Laplace method for multivariate phase-type distributions}
\usage{
\S4method{laplace}{mph}(x, r)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{r}{A matrix of real values.}
}
\value{
A vector containing the corresponding Laplace transform evaluations.
}
\description{
Laplace method for multivariate phase-type distributions
}
\examples{
set.seed(123)
obj <- mph(structure = c("general", "general"))
laplace(obj, matrix(c(0.5, 1), ncol = 2))
}
