% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{laplace}
\alias{laplace}
\title{New Generic for Laplace transform of Matrix Distributions}
\usage{
laplace(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
Laplace transform of the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
