% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.bivph.R
\name{linCom,bivph-method}
\alias{linCom,bivph-method}
\title{Linear Combination method for bivariate phase-type distributions}
\usage{
\S4method{linCom}{bivph}(x, w = c(1, 1))
}
\arguments{
\item{x}{An object of class \linkS4class{bivph}.}

\item{w}{A vector with non-negative entries.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Linear Combination method for bivariate phase-type distributions
}
\examples{
obj <- bivph(dimensions = c(3, 3))
linCom(obj, c(1, 0))
}
