% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{linCom,MPHstar-method}
\alias{linCom,MPHstar-method}
\title{Linear combination method for MPHstar class}
\usage{
\S4method{linCom}{MPHstar}(x, w)
}
\arguments{
\item{x}{An object of class \linkS4class{MPHstar}.}

\item{w}{A vector with non-negative entries.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Linear combination method for MPHstar class
}
\examples{
obj <- MPHstar(structure = "general")
linCom(obj, c(1, 0))
}
