% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linear_combination}
\alias{linear_combination}
\title{Pi and T of a linear combination of a MPH*}
\usage{
linear_combination(w, pi, T, R)
}
\description{
Pi and T of a linear combination of a MPH*
}
\examples{
pi <- c(0.15, 0.85, 0 ,0)
T11 <- matrix(c(c(-2,9),c(0,-11)), nrow = 2, ncol = 2)
T12 <- matrix(c(c(2,0),c(0,2)), nrow = 2, ncol = 2)
T22 <- matrix(c(c(-1,0),c(0.5,-5)), nrow = 2, ncol = 2)
T <- merge_matrices(T11, T12, T22)
R <- matrix(c(c(1,1,0,0), c(0,0,1,1)), ncol=2)
w1 <- c(1,0)
linear_combination(w1, pi, T, R)
w2 <- c(0,1)
linear_combination(w2, pi, T, R)
matrix(c(0.15, 0.85), ncol=2)\%*\%matrix_inverse(T11 * (-1))\%*\%T12
w3 <- c(1,1)
linear_combination(w3, pi, T, R)
}
