% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LInf_norm}
\alias{LInf_norm}
\title{Returns the  L-oo norm of a matrix}
\usage{
LInf_norm(A)
}
\description{
The matrix L-oo norm is defined as:
L-oo A =  max ( 1 <= I <= M ) sum ( 1 <= J <= N ) abs ( A(I,J) ).
}
