% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LInf_normArma_0}
\alias{LInf_normArma_0}
\title{L inf norm of a matrix}
\usage{
LInf_normArma_0(A)
}
\arguments{
\item{A}{A matrix.}
}
\description{
Computes the L inf norm of a matrix \code{A}, which is defined as:
L_inf A =  max ( 1 <= I <= M ) sum ( 1 <= J <= N ) abs ( A(I,J) ).
}
