% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixt.R
\name{lmvgamma}
\alias{lmvgamma}
\title{lmvgamma}
\usage{
lmvgamma(x, p)
}
\arguments{
\item{x}{non-negative numeric vector, matrix, or array}

\item{p}{positive integer, dimension of a square matrix}
}
\value{
log of multivariate gamma for each entry of x. For non-log variant, see mvgamma.
}
\description{
A special mathematical function related to the gamma function,
    generalized for multivariate gammas.
}
\examples{
lgamma(1:12)
lmvgamma(1:12,1)
}
\seealso{
mvgamma
}
