% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihood_UNIs_PI}
\alias{logLikelihood_UNIs_PI}
\title{Loglikelihood of PI using uniformization, with customizable observation transformation, automatically handling right-censored or interval-censored data based on the structure of \code{rcens}.}
\usage{
logLikelihood_UNIs_PI(
  h,
  alpha,
  S,
  beta1,
  beta2,
  obs,
  weight,
  rcens,
  rcweight,
  scale1,
  scale2,
  gfun_name
)
}
\arguments{
\item{h}{Positive parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{beta1}{Parameter of transformation (uncensored observations).}

\item{beta2}{Parameter of transformation (censored observations).}

\item{obs}{The observations.}

\item{weight}{Weights of the observations.}

\item{rcens}{Censored observations (either numeric vector for right censoring or matrix for interval censoring).}

\item{rcweight}{Weights of the censored observations.}

\item{scale1}{Scale for observations.}

\item{scale2}{Scale for censored observations.}

\item{gfun_name}{Name of the transformation function to be applied to observations (e.g., "scale", "log", "sqrt").}
}
\value{
The loglikelihood of censored the sample.
}
\description{
Loglikelihood for a sample.
}
