% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodbivDPH_MoE}
\alias{logLikelihoodbivDPH_MoE}
\title{Loglikelihood for bivariate discrete phase-type MoE}
\usage{
logLikelihoodbivDPH_MoE(alpha, S11, S12, S22, obs, weight)
}
\arguments{
\item{alpha}{Initial probabilities.}

\item{S11}{Sub-transition matrix.}

\item{S12}{Matrix.}

\item{S22}{Sub-transition matrix.}

\item{obs}{The observations.}

\item{weight}{The weights of the observations.}
}
\description{
Loglikelihood for bivariate discrete phase-type MoE
}
