% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodbivPH}
\alias{logLikelihoodbivPH}
\title{Loglikelihood for Bivariate PH}
\usage{
logLikelihoodbivPH(alpha, S11, S12, S22, obs, weight)
}
\arguments{
\item{alpha}{Vector of initial probabilities.}

\item{S11}{Sub-intensity matrix.}

\item{S12}{Matrix.}

\item{S22}{Sub-intensity matrix.}

\item{obs}{The observations.}

\item{weight}{The weights of the observations.}
}
\description{
Loglikelihood for Bivariate PH
}
