% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodDPH}
\alias{logLikelihoodDPH}
\title{Loglikelihood for a sample}
\usage{
logLikelihoodDPH(alpha, S, obs, weight)
}
\arguments{
\item{alpha}{initial probabilities}

\item{S}{sub-intensity}

\item{obs}{the observations}

\item{weight}{weight of the observations}
}
\description{
Loglikelihood for a sample
}
