% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodDPH_MoE}
\alias{logLikelihoodDPH_MoE}
\title{Loglikelihood for discrete phase type MoE}
\usage{
logLikelihoodDPH_MoE(alpha, S, obs, weight)
}
\arguments{
\item{alpha}{initial probabilities}

\item{S}{sub-intensity}

\item{obs}{the observations}

\item{weight}{weight of the observations}
}
\description{
Loglikelihood for discrete phase type MoE
}
