% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodIPH_RK}
\alias{logLikelihoodIPH_RK}
\title{Loglikelihood IPH using RK and g as an input
One needs to be careful with the GEV since it is decreasing
It is slower than using the density directly - Perhaps is the iteration with R}
\usage{
logLikelihoodIPH_RK(
  h,
  pi,
  T,
  g,
  g_inv,
  lambda,
  beta,
  obs,
  weight,
  rcens,
  rcweight
)
}
\description{
Loglikelihood IPH using RK and g as an input
One needs to be careful with the GEV since it is decreasing
It is slower than using the density directly - Perhaps is the iteration with R
}
