% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodmDPH}
\alias{logLikelihoodmDPH}
\title{Loglikelihood for multivariate discrete phase-type}
\usage{
logLikelihoodmDPH(alpha, S_list, obs, weight)
}
\arguments{
\item{alpha}{Initial probabilities.}

\item{S_list}{List of marginal sub-transition matrices.}

\item{obs}{The observations.}

\item{weight}{The weights of the observations.}
}
\description{
Loglikelihood for multivariate discrete phase-type
}
