% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMGEV_RKs1}
\alias{logLikelihoodMGEV_RKs1}
\title{Loglikelihood of matrix GEV using RK
I am assuming that the sample is given in an increasing order}
\usage{
logLikelihoodMGEV_RKs1(
  h,
  pi,
  T,
  beta,
  obs,
  weight,
  rcens,
  rcweight,
  scale1,
  scale2
)
}
\description{
Loglikelihood of matrix GEV using RK
I am assuming that the sample is given in an increasing order
}
