% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMGEVD_RK}
\alias{logLikelihoodMGEVD_RK}
\title{Loglikelihood of matrix GEV using RK
I am assuming that the sample is given in an increasing order}
\usage{
logLikelihoodMGEVD_RK(h, pi, T, beta, obs, weight, rcens, rcweight)
}
\description{
Loglikelihood of matrix GEV using RK
I am assuming that the sample is given in an increasing order
}
