% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMGomp_RK}
\alias{logLikelihoodMGomp_RK}
\title{Loglikelihood of matrix Gompertz using RK}
\usage{
logLikelihoodMGomp_RK(h, pi, T, beta, obs, weight, rcens, rcweight)
}
\description{
Loglikelihood of matrix Gompertz using RK
}
