% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMgompertz_RKs}
\alias{logLikelihoodMgompertz_RKs}
\title{Loglikelihood of matrix-Gompertz using RK}
\usage{
logLikelihoodMgompertz_RKs(
  h,
  alpha,
  S,
  beta,
  obs,
  weight,
  rcens,
  rcweight,
  scale1,
  scale2
)
}
\arguments{
\item{h}{Step-length.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity.}

\item{beta}{Parameter of transformation}

\item{obs}{The observations.}

\item{weight}{Weight of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{Weight of the censored observations.}

\item{scale1}{Scale for observations.}

\item{scale2}{Scale for censored observations.}
}
\description{
Loglikelihood for a sample.
}
