% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMloglogistic_PADEs}
\alias{logLikelihoodMloglogistic_PADEs}
\title{Loglikelihood of matrix-loglogistic using Pade}
\usage{
logLikelihoodMloglogistic_PADEs(
  h,
  alpha,
  S,
  beta,
  obs,
  weight,
  rcens,
  rcweight,
  scale1,
  scale2
)
}
\arguments{
\item{h}{Nuisance parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity.}

\item{beta}{Inhomogeneity parameter.}

\item{obs}{The observations.}

\item{weight}{Weight of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{Weight of the censored observations.}

\item{scale1}{Scale for observations.}

\item{scale2}{Scale for censored observations.}
}
\description{
Loglikelihood for a sample.
}
