% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMLogLogistic_RK}
\alias{logLikelihoodMLogLogistic_RK}
\title{Loglikelihood of matrix Log-Logistic using RK}
\usage{
logLikelihoodMLogLogistic_RK(h, pi, T, beta, obs, weight, rcens, rcweight)
}
\description{
Loglikelihood of matrix Log-Logistic using RK
}
