% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMLogLogistic_RKs1}
\alias{logLikelihoodMLogLogistic_RKs1}
\title{Loglikelihood of matrix Log-Logistic using RK}
\usage{
logLikelihoodMLogLogistic_RKs1(
  h,
  pi,
  T,
  beta,
  obs,
  weight,
  rcens,
  rcweight,
  scale1,
  scale2
)
}
\description{
Loglikelihood of matrix Log-Logistic using RK
}
