% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMLogNormal_RKs}
\alias{logLikelihoodMLogNormal_RKs}
\title{Loglikelihood of matrix LogNormal using RK
This is the fastest option}
\usage{
logLikelihoodMLogNormal_RKs(
  h,
  pi,
  T,
  beta,
  obs,
  weight,
  rcens,
  rcweight,
  scale1,
  scale2
)
}
\description{
Loglikelihood of matrix LogNormal using RK
This is the fastest option
}
