% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMPar_RK}
\alias{logLikelihoodMPar_RK}
\title{Loglikelihood of matrix Pareto using RK}
\usage{
logLikelihoodMPar_RK(h, pi, T, beta, obs, weight, rcens, rcweight)
}
\description{
Loglikelihood of matrix Pareto using RK
}
