% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMPar_RKs1}
\alias{logLikelihoodMPar_RKs1}
\title{Loglikelihood of matrix Pareto using RK}
\usage{
logLikelihoodMPar_RKs1(
  h,
  pi,
  T,
  beta,
  obs,
  weight,
  rcens,
  rcweight,
  scale1,
  scale2
)
}
\description{
Loglikelihood of matrix Pareto using RK
}
