% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMpareto_UNI_inhom_intCens}
\alias{logLikelihoodMpareto_UNI_inhom_intCens}
\title{Loglikelihood of matrix-Pareto using uniformization, when the intensity function is regressed on covariate informatino}
\usage{
logLikelihoodMpareto_UNI_inhom_intCens(
  h,
  alpha,
  S,
  beta,
  obs,
  weight,
  rcens,
  rcweight
)
}
\arguments{
\item{h}{Positive parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{beta}{Parameter of transformation.}

\item{obs}{The observations.}

\item{weight}{Weights of the observations.}

\item{rcens}{Interval censored observations.}

\item{rcweight}{Weights of the censored observations.}
}
\description{
Loglikelihood for a sample.
}
