% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMweibull_PADE}
\alias{logLikelihoodMweibull_PADE}
\title{Loglikelihood of matrix-Weibull using Pade}
\usage{
logLikelihoodMweibull_PADE(h, alpha, S, beta, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Nuisance parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity.}

\item{beta}{Inhomogeneity parameter.}

\item{obs}{The observations.}

\item{weight}{Weight of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{Weight of the censored observations.}
}
\description{
Loglikelihood for a sample.
}
