% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMWeibull_RK}
\alias{logLikelihoodMWeibull_RK}
\title{Loglikelihood of matrix Weibull using RK
This is the fastest option}
\usage{
logLikelihoodMWeibull_RK(h, pi, T, beta, obs, weight, rcens, rcweight)
}
\description{
Loglikelihood of matrix Weibull using RK
This is the fastest option
}
