% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_MoE}
\alias{logLikelihoodPH_MoE}
\title{Loglikelihood for a sample}
\usage{
logLikelihoodPH_MoE(alpha1, alpha2, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{alpha1}{initial probabilities non-censored data}

\item{alpha2}{initial probabilities censored data}

\item{S}{sub-intensity}

\item{obs}{the observations}

\item{weight}{weight of the observations}

\item{rcens}{censored observations}

\item{rcweight}{weight of the censored observations}
}
\description{
Loglikelihood for a sample
}
