% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_PADE}
\alias{logLikelihoodPH_PADE}
\title{Loglikelihood of PH using Pade}
\usage{
logLikelihoodPH_PADE(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Nuisance parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity.}

\item{obs}{The observations.}

\item{weight}{Weight of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{Weight of the censored observations.}
}
\description{
Loglikelihood for a sample.
}
