% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_RK_old}
\alias{logLikelihoodPH_RK_old}
\title{Loglikelihood using RK}
\usage{
logLikelihoodPH_RK_old(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Step-length.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity.}

\item{obs}{The observations.}

\item{weight}{Weight of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{Weight of the censored observations.}
}
\description{
Loglikelihood for a sample.
}
