% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_RKs}
\alias{logLikelihoodPH_RKs}
\title{Loglikelihood using RK, with scale}
\usage{
logLikelihoodPH_RKs(h, pi, T, obs, weight, rcens, rcweight, scale1, scale2)
}
\description{
Loglikelihood using RK, with scale
}
