% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_UNI}
\alias{logLikelihoodPH_UNI}
\title{Loglikelihood using Uniformization}
\usage{
logLikelihoodPH_UNI(epsilon, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{epsilon}{positive parameter}

\item{alpha}{initial probabilities}

\item{S}{sub-intensity}

\item{obs}{the observations}

\item{weight}{weight of the observations}

\item{rcens}{censored observations}

\item{rcweight}{weight of the censored observations}
}
\description{
Loglikelihood for a sample
}
