% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_UNI_intervalCensoring}
\alias{logLikelihoodPH_UNI_intervalCensoring}
\title{Loglikelihood of phase-type using uniformization}
\usage{
logLikelihoodPH_UNI_intervalCensoring(
  h,
  alpha,
  S,
  obs,
  weight,
  rcens,
  rcweight
)
}
\arguments{
\item{h}{Positive parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{obs}{The observations.}

\item{weight}{Weights of the observations.}

\item{rcens}{Interval censored observations.}

\item{rcweight}{Weights of the censored observations.}
}
\description{
Loglikelihood for a sample.
}
