% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{LRT,ph,ph-method}
\alias{LRT,ph,ph-method}
\title{LRT method for ph class}
\usage{
\S4method{LRT}{ph,ph}(x, y)
}
\arguments{
\item{x, y}{Objects of class \linkS4class{ph}.}
}
\value{
LRT between the models.
}
\description{
LRT method for ph class
}
