% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{m_exp_sum}
\alias{m_exp_sum}
\title{Computes exp(Sx) base on the values on powerVector}
\usage{
m_exp_sum(x, n, powerVector, a)
}
\arguments{
\item{x}{A number.}

\item{n}{An integer.}

\item{powerVector}{A vector.}

\item{a}{A number.}
}
\description{
Computes exp(Sx) base on the values on powerVector
}
