% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{m_plot,ph-method}
\alias{m_plot,ph-method}
\title{Plot Method for phase type distributions}
\usage{
\S4method{m_plot}{ph}(x, y = NA)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{a dataset}
}
\description{
Plot Method for phase type distributions
}
\examples{

}
