% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{m_plot}
\alias{m_plot}
\title{New Generic for Plotting Matrix Distributions}
\usage{
m_plot(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{}

\item{y}{a vector of data.}
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
