% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4c.biviph.R
\name{marginal,biviph-method}
\alias{marginal,biviph-method}
\title{Marginal method for biviph class}
\usage{
\S4method{marginal}{biviph}(x, mar = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{biviph}.}

\item{mar}{Indicator of which marginal.}
}
\value{
An object of the of class \linkS4class{iph}.
}
\description{
Marginal method for biviph class
}
\examples{
under_bivph <- bivph(dimensions = c(3, 3))
obj <- biviph(under_bivph, gfun = c("weibull", "pareto"), gfun_pars = list(c(2), c(3)))
marginal(obj, 1)
}
