% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.bivph.R
\name{marginal,bivph-method}
\alias{marginal,bivph-method}
\title{Marginal method for bivph class}
\usage{
\S4method{marginal}{bivph}(x, mar = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{bivph}.}

\item{mar}{Indicator of which marginal.}
}
\value{
An object of the of class \linkS4class{ph}.
}
\description{
Marginal method for bivph class
}
\examples{
obj <- bivph(dimensions = c(3, 3))
marginal(obj, 1)
}
