% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b.miph.R
\name{marginal,miph-method}
\alias{marginal,miph-method}
\title{Marginal method for multivariate inhomogeneous phase-type distributions}
\usage{
\S4method{marginal}{miph}(x, mar = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{miph}.}

\item{mar}{Indicator of which marginal.}
}
\value{
An object of the of class \linkS4class{iph}.
}
\description{
Marginal method for multivariate inhomogeneous phase-type distributions
}
\examples{
under_mph <- mph(structure = c("general", "general"))
obj <- miph(under_mph, gfun = c("weibull", "pareto"), gfun_pars = list(c(2), c(3)))
marginal(obj, 1)
}
