% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{marginal,mph-method}
\alias{marginal,mph-method}
\title{Marginal method for multivariate phase-type distributions}
\usage{
\S4method{marginal}{mph}(x, mar = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{mar}{Indicator of which marginal.}
}
\value{
An object of the of class \linkS4class{ph}.
}
\description{
Marginal method for multivariate phase-type distributions
}
\examples{
obj <- mph(structure = c("general", "general"))
marginal(obj, 1)
}
