% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{marginal,MPHstar-method}
\alias{marginal,MPHstar-method}
\title{Marginal method for MPHstar class}
\usage{
\S4method{marginal}{MPHstar}(x, mar = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{MPHstar}.}

\item{mar}{Indicator of which marginal.}
}
\value{
An object of the of class \linkS4class{ph}.
}
\description{
Marginal method for MPHstar class
}
\examples{
obj <- MPHstar(structure = "general")
marginal(obj, 1)
}
