% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{marginal}
\alias{marginal}
\title{New generic for the marginals of multivariate matrix distributions}
\usage{
marginal(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
Marginal of the matrix distribution.
}
\description{
Methods are available for objects of multivariate classes.
}
