% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{marginal_expectation}
\alias{marginal_expectation}
\title{Marginal conditional expectations}
\usage{
marginal_expectation(rew, pos, N, alpha, S, obs, weight)
}
\arguments{
\item{rew}{Column of the reward matrix corresponding to its marginal.}

\item{pos}{Vector that indicates which state is associated to a positive reward.}

\item{N}{Uniformization parameter.}

\item{alpha}{Marginal initial distribution vector.}

\item{S}{Marginal sub-intensity matrix.}

\item{obs}{Marginal observations.}

\item{weight}{Marginal weights.}
}
\value{
A vector with the expected time spent in each state by the marginal, conditional on the observations.
}
\description{
Marginal conditional expectations
}
