% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_exponential_slow}
\alias{matrix_exponential_slow}
\title{Matrix exponential algorithm}
\usage{
matrix_exponential_slow(A)
}
\arguments{
\item{A}{a matrix}
}
\description{
MATLAB's built-in algorithm - Pade approximation
}
