% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_exponential_tem}
\alias{matrix_exponential_tem}
\title{Matrix exponential algorithm}
\usage{
matrix_exponential_tem(A)
}
\arguments{
\item{A}{A matrix.}
}
\description{
MATLAB's built-in algorithm - Pade approximation.
}
