% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_product}
\alias{matrix_product}
\title{Product of two matrices}
\usage{
matrix_product(A1, A2)
}
\arguments{
\item{A1}{Matrix.}

\item{A2}{Matrix.}
}
\value{
Computes C = A1 * A2.
}
\description{
Product of two matrices
}
