% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_vanloan}
\alias{matrix_vanloan}
\title{Creates the matrix  (A1, B1 ; 0, A2)}
\usage{
matrix_vanloan(A1, A2, B1)
}
\arguments{
\item{A1}{Matrix.}

\item{A2}{Matrix.}

\item{B1}{Matrix.}
}
\value{
Computes (A1, B1 ; 0, A2).
}
\description{
Creates the matrix  (A1, B1 ; 0, A2)
}
