% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_VanLoanArma}
\alias{matrix_VanLoanArma}
\title{Creates the matrix  (A1, B1 ; 0, A2)}
\usage{
matrix_VanLoanArma(A1, A2, B1)
}
\arguments{
\item{A1}{matrix}

\item{A2}{matrix}

\item{B1}{matrix}
}
\value{
Computes (A1, B1 ; 0, A2)
}
\description{
Creates the matrix  (A1, B1 ; 0, A2)
}
