% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixExpSum_arma}
\alias{matrixExpSum_arma}
\title{Computes e^(Sx) base on the values on powerVector}
\usage{
matrixExpSum_arma(x, n, powerVector, a)
}
\arguments{
\item{x}{a number}

\item{n}{an integer}

\item{powerVector}{a vector}

\item{a}{a number}
}
\description{
Computes e^(Sx) base on the values on powerVector
}
