% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrixMaxDiagonal}
\alias{matrixMaxDiagonal}
\title{Maximum entry in the diagonal of a matrix}
\usage{
matrixMaxDiagonal(A)
}
\arguments{
\item{A}{a matrix}
}
\description{
Maximum entry in the diagonal of a matrix
}
