% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{max_diagonal}
\alias{max_diagonal}
\title{Maximum diagonal element of a matrix}
\usage{
max_diagonal(A)
}
\arguments{
\item{A}{Matrix.}
}
\value{
The maximum value in the diagonal.
}
\description{
Maximum diagonal element of a matrix
}
