% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{maximum,iph,iph-method}
\alias{maximum,iph,iph-method}
\title{Maximum Method for inhomogeneous phase type distributions}
\usage{
\S4method{maximum}{iph,iph}(x1, x2)
}
\arguments{
\item{x1}{an object of class \linkS4class{iph}.}

\item{x2}{an object of class \linkS4class{iph}.}
}
\description{
Maximum Method for inhomogeneous phase type distributions
}
