% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{maximum,ph,ph-method}
\alias{maximum,ph,ph-method}
\title{Maximum Method for phase type distributions}
\usage{
\S4method{maximum}{ph,ph}(x1, x2)
}
\arguments{
\item{x1}{an object of class \linkS4class{ph}.}

\item{x2}{an object of class \linkS4class{ph}.}
}
\description{
Maximum Method for phase type distributions
}
