% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{maximum}
\alias{maximum}
\title{New Generic for Maximum of two Matrix Distributions}
\usage{
maximum(x1, x2, ...)
}
\arguments{
\item{...}{}

\item{x}{an object of the model class.}
}
\value{
A realization from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
