% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\docType{class}
\name{mdph-class}
\alias{mdph-class}
\title{Multivariate Discrete Phase Type distributions}
\value{
Class object.
}
\description{
Class of objects for multivariate discrete phase-type distributions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the discrete phase type distribution.}

\item{\code{pars}}{A list comprising of the parameters.}

\item{\code{fit}}{A list containing estimation information.}
}}

