% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{mdph}
\alias{mdph}
\title{Constructor Function for multivariate discrete phase-type distributions}
\usage{
mdph(alpha = NULL, S = NULL, structure = NULL, dimension = 3, variables = NULL)
}
\arguments{
\item{alpha}{A probability vector.}

\item{S}{A list of sub-transition matrices.}

\item{structure}{A vector of valid ph structures.}

\item{dimension}{The dimension of the dph structure (if provided).}

\item{variables}{The dimension of the multivariate discrete phase-type.}
}
\value{
An object of class \linkS4class{mdph}.
}
\description{
Constructor Function for multivariate discrete phase-type distributions
}
