% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\name{mean,bivdph-method}
\alias{mean,bivdph-method}
\title{Mean Method for bivdph class}
\usage{
\S4method{mean}{bivdph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{bivdph}.}
}
\value{
The mean of the bivariate discrete phase-type distribution.
}
\description{
Mean Method for bivdph class
}
\examples{
obj <- bivdph(dimensions = c(3, 3))
mean(obj)
}
