% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.bivph.R
\name{mean,bivph-method}
\alias{mean,bivph-method}
\title{Mean Method for bivph class}
\usage{
\S4method{mean}{bivph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{bivph}.}
}
\value{
The mean of the bivariate phase-type distribution.
}
\description{
Mean Method for bivph class
}
\examples{
obj <- bivph(dimensions = c(3, 3))
mean(obj)
}
