% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{mean,dph-method}
\alias{mean,dph-method}
\title{Mean Method for discrete phase-type distributions}
\usage{
\S4method{mean}{dph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{dph}.}
}
\value{
The raw first moment of the \linkS4class{dph} object.
}
\description{
Mean Method for discrete phase-type distributions
}
\examples{
set.seed(123)
dph1 <- dph(structure = "general", dimension = 3)
mean(dph1)
}
