% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{mean,mdph-method}
\alias{mean,mdph-method}
\title{Mean method for multivariate discrete phase-type distributions}
\usage{
\S4method{mean}{mdph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{mdph}.}
}
\value{
The mean of the multivariate discrete phase-type distribution.
}
\description{
Mean method for multivariate discrete phase-type distributions
}
\examples{
obj <- mdph(structure = c("general", "general"))
mean(obj)
}
