% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{mean,mph-method}
\alias{mean,mph-method}
\title{Mean method for multivariate phase-type distributions}
\usage{
\S4method{mean}{mph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}
}
\value{
The mean of the multivariate phase-type distribution.
}
\description{
Mean method for multivariate phase-type distributions
}
\examples{
obj <- mph(structure = c("general", "general"))
mean(obj)
}
