% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{mean,MPHstar-method}
\alias{mean,MPHstar-method}
\title{Mean method for MPHstar class}
\usage{
\S4method{mean}{MPHstar}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{MPHstar}.}
}
\value{
The mean of MPHstar distribution.
}
\description{
Mean method for MPHstar class
}
\examples{
obj <- MPHstar(structure = "general")
mean(obj)
}
