% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{mean,ph-method}
\alias{mean,ph-method}
\title{Mean Method for phase-type distributions}
\usage{
\S4method{mean}{ph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{ph}.}
}
\value{
The raw first moment of the \linkS4class{ph} (or undelying \linkS4class{ph}) object.
}
\description{
Mean Method for phase-type distributions
}
\examples{
set.seed(123)
ph1 <- ph(structure = "general", dimension = 3)
mean(ph1)
}
