% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_matrices}
\alias{merge_matrices}
\title{Merges the matrices T11, T12 and T22 of a bivariate matrix}
\usage{
merge_matrices(T11, T12, T22)
}
\description{
Merges the matrices T11, T12 and T22 of a bivariate matrix
}
