% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mexponentialArma}
\alias{mexponentialArma}
\title{Matrix exponential algorithm}
\usage{
mexponentialArma(Ainput)
}
\arguments{
\item{Ainput}{a matrix}
}
\description{
MATLAB's built-in algorithm - Pade approximation
}
