% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgevcdf}
\alias{mgevcdf}
\title{Matrix GEV cdf}
\usage{
mgevcdf(x, alpha, S, mu, sigma, xi, lower_tail = TRUE)
}
\arguments{
\item{x}{Non-negative value.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{xi}{Shape parameter.}

\item{lower_tail}{Cdf or tail.}
}
\value{
The cdf (tail) at \code{x}.
}
\description{
Computes the cdf (tail) of a matrix GEV distribution with parameters 
\code{alpha}, \code{S} and \code{beta} at \code{x}.
}
