% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.bivph.R
\name{mgf,bivph-method}
\alias{mgf,bivph-method}
\title{Mgf method for bivph class}
\usage{
\S4method{mgf}{bivph}(x, r)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{r}{A matrix of real values.}
}
\value{
A vector containing the corresponding mgf evaluations.
}
\description{
Mgf method for bivph class
}
\examples{
set.seed(123)
obj <- bivph(dimensions = c(3, 3))
mgf(obj, matrix(c(0.5, 0.1), ncol = 2))
}
