% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{mgf,mph-method}
\alias{mgf,mph-method}
\title{Mgf method for multivariate phase-type distributions}
\usage{
\S4method{mgf}{mph}(x, r)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{r}{A matrix of real values.}
}
\value{
A vector containing the corresponding mgf evaluations.
}
\description{
Mgf method for multivariate phase-type distributions
}
\examples{
set.seed(124)
obj <- mph(structure = c("general", "general"))
mgf(obj, matrix(c(0.5, 0.3), ncol = 2))
}
