% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{mgf,ph-method}
\alias{mgf,ph-method}
\title{Mgf method for phase-type distributions}
\usage{
\S4method{mgf}{ph}(x, r)
}
\arguments{
\item{x}{An object of class \linkS4class{ph}.}

\item{r}{A vector of real values.}
}
\value{
The mgf of the \linkS4class{ph} (or underlying \linkS4class{ph}) object
at the given locations.
}
\description{
Mgf method for phase-type distributions
}
\examples{
set.seed(123)
obj <- ph(structure = "general", dimension = 3)
mgf(obj, 0.4)
}
