% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{mgf}
\alias{mgf}
\title{New Generic for mgf of Matrix Distributions}
\usage{
mgf(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
Mgf transform of the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
