% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mGompertzden}
\alias{mGompertzden}
\title{Matrix Gompertz density}
\usage{
mGompertzden(x, pi, T, beta)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{parameter}
}
\value{
The density at \code{x}
}
\description{
Computes the density of a matrix Gompertz distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
}
\examples{
alpha <- c(0.5, 0.3, 0.2)
T <- matrix(c(c(-1,0,0),c(1,-2,0),c(0,1,-5)), nrow = 3, ncol = 3)
beta <- 0.5
mgompertzden(0.5, alpha, T, beta) 
}
