% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{minimum,dph,dph-method}
\alias{minimum,dph,dph-method}
\title{Minimum method for discrete phase-type distributions}
\usage{
\S4method{minimum}{dph,dph}(x1, x2)
}
\arguments{
\item{x1}{An object of class \linkS4class{dph}.}

\item{x2}{An object of class \linkS4class{dph}.}
}
\value{
An object of class \linkS4class{dph}.
}
\description{
Minimum method for discrete phase-type distributions
}
\examples{
dph1 <- dph(structure = "general", dimension = 3)
dph2 <- dph(structure = "general", dimension = 5)
dph_min <- minimum(dph1, dph2)
dph_min
}
