% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{minimum,iph,iph-method}
\alias{minimum,iph,iph-method}
\title{Minimum Method for inhomogeneous phase type distributions}
\usage{
\S4method{minimum}{iph,iph}(x1, x2)
}
\arguments{
\item{x1}{an object of class \linkS4class{iph}.}

\item{x2}{an object of class \linkS4class{iph}.}
}
\description{
Minimum Method for inhomogeneous phase type distributions
}
