% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{minimum,ph,ph-method}
\alias{minimum,ph,ph-method}
\title{Minimum Method for phase type distributions}
\usage{
\S4method{minimum}{ph,ph}(x1, x2)
}
\arguments{
\item{x1}{an object of class \linkS4class{ph}.}

\item{x2}{an object of class \linkS4class{ph}.}
}
\description{
Minimum Method for phase type distributions
}
