% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{minimum}
\alias{minimum}
\title{New Generic for Minimum of two Matrix Distributions}
\usage{
minimum(x1, x2, ...)
}
\arguments{
\item{...}{}

\item{x}{an object of the model class.}
}
\value{
A realization from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
