% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b.miph.R
\docType{class}
\name{miph-class}
\alias{miph-class}
\title{Multivariate inhomogeneous phase-type distributions}
\value{
Class object.
}
\description{
Class of objects for multivariate inhomogeneous phase-type distributions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the phase type distribution.}

\item{\code{gfun}}{A list comprising of the parameters.}

\item{\code{scale}}{Scale.}
}}

