% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{mixture,dph,dph-method}
\alias{mixture,dph,dph-method}
\title{Mixture Method for phase-type distributions}
\usage{
\S4method{mixture}{dph,dph}(x1, x2, prob)
}
\arguments{
\item{x1}{An object of class \linkS4class{dph}.}

\item{x2}{An object of class \linkS4class{dph}.}

\item{prob}{Probability for first object.}
}
\value{
An object of class \linkS4class{dph}.
}
\description{
Mixture Method for phase-type distributions
}
\examples{
dph1 <- dph(structure = "general", dimension = 3)
dph2 <- dph(structure = "general", dimension = 5)
dph_mix <- mixture(dph1, dph2, 0.5)
dph_mix
}
