% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{mixture,ph,ph-method}
\alias{mixture,ph,ph-method}
\title{Mixture Method for phase-type distributions}
\usage{
\S4method{mixture}{ph,ph}(x1, x2, prob)
}
\arguments{
\item{x1}{An object of class \linkS4class{ph}.}

\item{x2}{An object of class \linkS4class{ph}.}

\item{prob}{Probability for first object.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Mixture Method for phase-type distributions
}
\examples{
ph1 <- ph(structure = "general", dimension = 3)
ph2 <- ph(structure = "gcoxian", dimension = 5)
ph_mix <- mixture(ph1, ph2, 0.5)
ph_mix
}
