% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixnorm.R
\name{mle.matrixnorm}
\alias{mle.matrixnorm}
\title{mle.matrixnorm}
\usage{
mle.matrixnorm(data, row.restrict = "none", col.restrict = "none",
  tol = 1e-09, max.iter = 100, U, V)
}
\arguments{
\item{data}{Either a list of matrices or a 3-D array with matrices in dimensions 2 and 3, indexed by dimension 1.}

\item{row.restrict}{At present, no options available for this. In the future, intend to have options for, eg, AR(1) structure.}

\item{col.restrict}{At present, no options available for this. In the future, intend to have options for, eg, AR(1) structure.}

\item{tol}{Convergence criterion. Measured against square deviation between iterations of the two variance-covariance matrices.}

\item{max.iter}{Maximum possible iterations of the algorithm.}

\item{U}{(optional) Can provide a starting point for the U matrix. By default, an identity matrix.}

\item{V}{(optional) Can provide a starting point for the V matrix. By default, an identity matrix.}
}
\value{
Returns a list with a mean matrix, a $U$ matrix, a $V$ matrix, the number of iterations, and error at the time of stopping.
}
\description{
mle.matrixnorm
}
\examples{
set.seed(20180202)
A = rmatrixnorm(n=100,mean=matrix(c(100,0,-100,0,25,-1000),nrow=2),
   L=matrix(c(2,1,0,.1),nrow=2),list=TRUE)
results=mle.matrixnorm(A)
print(results)


}
