% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mloglogisticcdf}
\alias{mloglogisticcdf}
\title{Matrix Log-Logistic cdf}
\usage{
mloglogisticcdf(x, pi, T, beta, lower_tail = TRUE)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{shape parameter}

\item{lower_tail}{cdf or tail}
}
\value{
The cdf (tail) at \code{x}
}
\description{
Computes the cdf (tail) of a matrix Log-Logistic distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
}
