% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mLogLogisticden}
\alias{mLogLogisticden}
\title{Matrix Log-Logistic density}
\usage{
mLogLogisticden(x, pi, T, beta)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{scale parameter}
}
\value{
The density at \code{x}
}
\description{
Computes the density of a matrix Log-Logistic distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
}
