% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\name{MoE,bivdph-method}
\alias{MoE,bivdph-method}
\title{MoE Method for bivdph Class}
\usage{
\S4method{MoE}{bivdph}(
  x,
  formula,
  data,
  alpha_vecs = NULL,
  weight = numeric(0),
  stepsEM = 1000,
  every = 10,
  rand_init = TRUE
)
}
\arguments{
\item{x}{An object of class \linkS4class{bivdph}.}

\item{formula}{A regression formula.}

\item{data}{A data frame.}

\item{alpha_vecs}{Matrix of initial probabilities.s}

\item{weight}{Vector of weights.}

\item{stepsEM}{Number of EM steps to be performed.}

\item{every}{Number of iterations between likelihood display updates.}

\item{rand_init}{Random initiation in the R-step.}
}
\value{
An object of class \linkS4class{sph}.
}
\description{
MoE Method for bivdph Class
}
