% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.phMoE.R
\name{MoE,ph-method}
\alias{MoE,ph-method}
\title{MoE Method for ph Class}
\usage{
\S4method{MoE}{ph}(
  x,
  formula,
  data,
  inhom = NULL,
  alpha_vecs = NULL,
  weight = numeric(0),
  delta = numeric(0),
  stepsEM = 1000,
  optim_method = "BFGS",
  maxit = 50,
  reltol = 1e-08,
  every = 10,
  rand_init = TRUE
)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{formula}{a regression formula.}

\item{data}{a data frame.}

\item{inhom}{a list with the inhomogeneity functions.}

\item{alpha_vecs}{matrix of initial probabilities.s}

\item{weight}{vector of weights.}

\item{delta}{right-censoring indicator.}

\item{stepsEM}{number of EM steps to be performed.}

\item{optim_method}{method to use in gradient optimization}

\item{maxit}{maximum number of iterations when optimizing g function.}

\item{reltol}{relative tolerance when optimizing g function.}

\item{every}{number of iterations between likelihood display updates.}

\item{rand_init}{random initiation in the R-step.}
}
\value{
An object of class \linkS4class{sph}.
}
\description{
MoE Method for ph Class
}
