% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\name{moment,bivdph-method}
\alias{moment,bivdph-method}
\title{Moment method for bivdph class}
\usage{
\S4method{moment}{bivdph}(x, k = c(1, 1))
}
\arguments{
\item{x}{An object of class \linkS4class{bivdph}.}

\item{k}{A vector with the location.}
}
\value{
An real value.
}
\description{
Moment method for bivdph class
}
\examples{
obj <- bivdph(dimensions = c(3, 3))
moment(obj, c(1, 1))
}
