% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.bivph.R
\name{moment,bivph-method}
\alias{moment,bivph-method}
\title{Moment method for bivph class}
\usage{
\S4method{moment}{bivph}(x, k = c(1, 1))
}
\arguments{
\item{x}{An object of class \linkS4class{bivph}.}

\item{k}{A vector with the location.}
}
\value{
An real value.
}
\description{
Moment method for bivph class
}
\examples{
obj <- bivph(dimensions = c(3, 3))
moment(obj, c(1, 1))
}
