% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{moment,dph-method}
\alias{moment,dph-method}
\title{Moment method for discrete phase-type distributions}
\usage{
\S4method{moment}{dph}(x, k = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{dph}.}

\item{k}{A positive integer (moment order).}
}
\value{
The factional moment of the \linkS4class{dph} object.
}
\description{
Moment method for discrete phase-type distributions
}
\examples{
set.seed(123)
obj <- dph(structure = "general", dimension = 3)
moment(obj, 2)
}
