% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{moment,mdph-method}
\alias{moment,mdph-method}
\title{Moment method for multivariate discrete phase-type distributions}
\usage{
\S4method{moment}{mdph}(x, k)
}
\arguments{
\item{x}{An object of class \linkS4class{mdph}.}

\item{k}{A vector of positive integer values.}
}
\value{
The corresponding joint factorial moment evaluation.
}
\description{
Moment method for multivariate discrete phase-type distributions
}
\examples{
obj <- mdph(structure = c("general", "general"))
moment(obj, c(2, 1))
}
