% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{moment,mph-method}
\alias{moment,mph-method}
\title{Moment method for multivariate phase-type distributions}
\usage{
\S4method{moment}{mph}(x, k)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{k}{A vector of non-negative integer values.}
}
\value{
The corresponding joint moment evaluation.
}
\description{
Moment method for multivariate phase-type distributions
}
\examples{
obj <- mph(structure = c("general", "general"))
moment(obj, c(2, 1))
}
