% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{moment,ph-method}
\alias{moment,ph-method}
\title{Moment Method for phase type distributions}
\usage{
\S4method{moment}{ph}(x, k = 1)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{k}{a positive integer.}
}
\description{
Moment Method for phase type distributions
}
