% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{moment}
\alias{moment}
\title{New generic for moments of matrix distributions}
\usage{
moment(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
Moment of the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
