% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\docType{class}
\name{mph-class}
\alias{mph-class}
\title{Multivariate Phase Type distributions}
\description{
Class of objects for multivariate phase type distributions
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the phase type distribution}

\item{\code{pars}}{a list comprising of the parameters.}
}}

