% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{mph}
\alias{mph}
\title{Constructor Function for multivariate phase type distributions}
\usage{
mph(
  ph = NULL,
  R = NULL,
  alpha = NULL,
  S = NULL,
  structure = NULL,
  dimension = 3
)
}
\arguments{
\item{R}{a compatible (non-negative) reward matrix}

\item{alpha}{a probability vector.}

\item{S}{a sub-intensity matrix.}

\item{structure}{a valid ph structure}

\item{dimension}{the dimension of the ph structure (if provided)}
}
\value{
An object of class \linkS4class{mph}.
}
\description{
Constructor Function for multivariate phase type distributions
}
