% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\docType{class}
\name{MPHstar-class}
\alias{MPHstar-class}
\title{Multivariate phase-type distributions obtained by transformation via rewards}
\description{
Class of objects for multivariate phase type distributions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the phase type distribution.}

\item{\code{pars}}{A list comprising of the parameters.}
}}

