% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{MPHstar}
\alias{MPHstar}
\title{Constructor Function for multivariate phase type distributions (MPH* class)}
\usage{
MPHstar(
  alpha = NULL,
  S = NULL,
  structure = NULL,
  dimension = 3,
  R = NULL,
  variables = 2
)
}
\arguments{
\item{alpha}{A probability vector.}

\item{S}{A sub-intensity matrix.}

\item{structure}{A valid ph structure.}

\item{dimension}{The dimension of the ph structure (if provided).}

\item{R}{A compatible (non-negative) reward matrix.}

\item{variables}{The number of desired marginals.}
}
\value{
An object of class \linkS4class{MPHstar}.
}
\description{
Constructor Function for multivariate phase type distributions (MPH* class)
}
\examples{
MPHstar(structure = "general", dimension = 4, variables = 3)
}
