% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{MPHstar_data_aggregation}
\alias{MPHstar_data_aggregation}
\title{Prepare data for the MPHstar_EMstep_UNI}
\usage{
MPHstar_data_aggregation(y, w = numeric(0))
}
\arguments{
\item{y}{A matrix with marginal observations, each column corresponds to a marginal.}

\item{w}{A matrix of weights, each column corresponds to a marginal.}
}
\value{
For summed and marginal observations we have a list with matrices of
 unique observations and their associated weights, separated by uncensored and right-censored data.
}
\description{
Prepare data for the MPHstar_EMstep_UNI
}
