% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MPHstar_EMstep_UNI}
\alias{MPHstar_EMstep_UNI}
\title{EM step using Uniformization for MPHstar class}
\usage{
MPHstar_EMstep_UNI(h, Rtol, alpha, S, R, mph_obs)
}
\arguments{
\item{h}{positive parameter for precision of uniformization method.}

\item{Rtol}{The smallest value that a reward can take.}

\item{alpha}{Vector of initial probabilities of the originating distribution.}

\item{S}{The sub-intensity matrix of the originating distribution.}

\item{R}{The reward matrix.}

\item{mph_obs}{The list of summed, marginal observations  with associated weights.}
}
\description{
EM step using Uniformization for MPHstar class
}
