% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_null}
\alias{n_null}
\title{Find how many states have null reward}
\usage{
n_null(R)
}
\arguments{
\item{R}{reward vector}
}
\value{
The number of states with null rewards
}
\description{
Find how many states have null reward
}
