% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{n_pos}
\alias{n_pos}
\title{Find how many states have positive reward}
\usage{
n_pos(R)
}
\arguments{
\item{R}{reward vector}
}
\value{
The number of states with positive rewards
}
\description{
Find how many states have positive reward
}
