% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_exit_vec}
\alias{new_exit_vec}
\title{Obtain new exit rate vector for the new embedded MC}
\usage{
new_exit_vec(R, Qtilda, S)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}

\item{S}{original sub-intensity matrix}
}
\value{
The exit rate vector for the new Markov Jump Process
}
\description{
Obtain new exit rate vector for the new embedded MC
}
