% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_pi}
\alias{new_pi}
\title{Obtain new sub-transition matrix for the new embedded MC}
\usage{
new_pi(R, Qtilda, alpha)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}

\item{alpha}{the original initial distribution vector}
}
\value{
The initial distribution vector for the new Markov Jump Process
}
\description{
Obtain new sub-transition matrix for the new embedded MC
}
