% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_state}
\alias{new_state}
\title{New state in a Markov jump process}
\usage{
new_state(prev_state, cum_embedded_mc, u)
}
\arguments{
\item{prev_state}{Previous state of the Markov jump process.}

\item{cum_embedded_mc}{Transition matrix.}

\item{u}{Random value in (0,1).}
}
\value{
Next state of the Markov jump process.
}
\description{
Given a transition matrix \code{Q}, a uniform value \code{u}, and a previous
 state \code{k}, it returns the new state of a Markov jump process.
}
