% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_subint_mat}
\alias{new_subint_mat}
\title{Obtain new sub-intensity matrix for the new MJP}
\usage{
new_subint_mat(R, Qtilda, S)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}

\item{S}{original sub-intensity matrix}
}
\value{
The sub-intensity matrix for the new Markov Jump Process
}
\description{
Obtain new sub-intensity matrix for the new MJP
}
