% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_trans_exit}
\alias{new_trans_exit}
\title{Obtain new sub-transition exit vector for the new embedded MC}
\usage{
new_trans_exit(R, Qtilda)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}
}
\value{
The exit rates for the new embedded Markov Chain
}
\description{
Obtain new sub-transition exit vector for the new embedded MC
}
