% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{new_trans_mat}
\alias{new_trans_mat}
\title{Obtain new sub-transition matrix for the new embedded MC}
\usage{
new_trans_mat(R, Qtilda)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}
}
\value{
The sub-transition matrix of the new embedded Markov Chain
}
\description{
Obtain new sub-transition matrix for the new embedded MC
}
