% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newState}
\alias{newState}
\title{New state in a Markov jump process}
\usage{
newState(previousState, cumulatedEmbeddedMC, u)
}
\arguments{
\item{previousState}{Previous state of the Markov jump process}

\item{cumulatedEmbeddedMC}{A transition matrix}

\item{u}{A random value in (0,1)}
}
\value{
The next state of the Markov jump process
}
\description{
Given a transition matrix \code{Q}, a uniform value \code{u}, and a previous state \code{k}, it returns the new state of a Markov jump process
}
