% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{Nfold,dph-method}
\alias{Nfold,dph-method}
\title{Nfold method for phase-type distributions}
\usage{
\S4method{Nfold}{dph}(x1, x2)
}
\arguments{
\item{x1}{An object of class \linkS4class{ph}.}

\item{x2}{An object of class \linkS4class{dph}.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Nfold method for phase-type distributions
}
\examples{
dph1 <- dph(structure = "general", dimension = 3)
dph2 <- dph(structure = "general", dimension = 2)
ph0 <- ph(structure = "general", dimension = 2)
Nfold(dph1, ph0)
Nfold(dph1, dph2)
}
