% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{Nfold}
\alias{Nfold}
\title{New generic for N-fold convolution of two matrix distributions}
\usage{
Nfold(x1, x2, ...)
}
\arguments{
\item{x1}{An object of the class \linkS4class{dph}.}

\item{x2}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
An object of the model class.
}
\description{
Methods are available for objects of classes \linkS4class{ph} and \linkS4class{dph}.
}
