% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b.miph.nnet.R
\name{nnet_fit,mph-method}
\alias{nnet_fit,mph-method}
\title{Fit Method for mph Class, using nueral network}
\usage{
\S4method{nnet_fit}{mph}(
  x,
  formula,
  y,
  data,
  alpha_mat = NULL,
  delta = numeric(0),
  stepsEM = 1000,
  r = 1,
  maxit = 100,
  reltol = 1e-08,
  rand_init = T
)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{formula}{a regression formula}

\item{y}{A matrix of observations.}

\item{data}{A data frame.}

\item{alpha_mat}{Matrix with initial distribution vectors for each row of observations}

\item{delta}{Matrix with right-censoring indicators. (1 uncensored, 0 right censored)}

\item{stepsEM}{Number of EM steps to be performed.}

\item{r}{Sub-sampling parameter, defaults to 1.}

\item{maxit}{Maximum number of iterations when optimizing g function.}

\item{reltol}{Relative tolerance when optimizing g function.}

\item{rand_init}{random initiation in the R-step.}
}
\description{
Fit Method for mph Class, using nueral network
}
