% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{nnet_fit}
\alias{nnet_fit}
\title{New Generic for Estimating Matrix Distributions with neural networks}
\usage{
nnet_fit(x, y, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{y}{A Matrix of data.}

\item{...}{Further parameters to be passed on.}
}
\value{
An object of the fitted model class.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
