% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{null_states}
\alias{null_states}
\title{Find which states have null reward}
\usage{
null_states(R)
}
\arguments{
\item{R}{reward vector}
}
\value{
A vector with the states (number) that are associated with null rewards
}
\description{
Find which states have null reward
}
