% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\docType{methods}
\name{p,ph-method}
\alias{p,ph-method}
\title{Distribution Method for phase type distributions}
\usage{
\S4method{p}{ph}(x, q = seq(0, 5, length.out = 100))
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{locations}
}
\value{
Density evaluated at locations
}
\description{
Distribution Method for phase type distributions
}
\examples{

}
