% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{p}
\alias{p}
\title{New Generic for the Distribution of Matrix Distributions}
\usage{
p(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{}
}
\value{
CDF from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
