% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\name{pgf,bivdph-method}
\alias{pgf,bivdph-method}
\title{Pgf method for bivariate discrete phase-type distributions}
\usage{
\S4method{pgf}{bivdph}(x, z)
}
\arguments{
\item{x}{An object of class \linkS4class{bivdph}.}

\item{z}{A vector of real values.}
}
\value{
The joint pdf of the \linkS4class{dph} object at the
given location.
}
\description{
Pgf method for bivariate discrete phase-type distributions
}
\examples{
obj <- bivdph(dimensions = c(3, 3))
pgf(obj, c(0.5, 0.2))
}
