% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{pgf,dph-method}
\alias{pgf,dph-method}
\title{Pgf Method for discrete phase-type distributions}
\usage{
\S4method{pgf}{dph}(x, z)
}
\arguments{
\item{x}{An object of class \linkS4class{dph}.}

\item{z}{A vector of real values.}
}
\value{
The probability generating of the \linkS4class{dph} object at the
 given locations.
}
\description{
Pgf Method for discrete phase-type distributions
}
\examples{
set.seed(123)
obj <- dph(structure = "general", dimension = 3)
pgf(obj, 0.5)
}
