% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{pgf,mdph-method}
\alias{pgf,mdph-method}
\title{Pgf method for multivariate discrete phase-type distributions}
\usage{
\S4method{pgf}{mdph}(x, z)
}
\arguments{
\item{x}{An object of class \linkS4class{mdph}.}

\item{z}{A matrix of real values.}
}
\value{
A vector containing the corresponding pgf evaluations.
}
\description{
Pgf method for multivariate discrete phase-type distributions
}
\examples{
obj <- mdph(structure = c("general", "general"))
pgf(obj, matrix(c(0.5, 1), ncol = 2))
}
