% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ph_laplace}
\alias{ph_laplace}
\title{Laplace transform of a phase-type distribution}
\usage{
ph_laplace(r, alpha, S)
}
\arguments{
\item{r}{Vector of real values.}

\item{alpha}{Vector of initial probabilities.}

\item{S}{Sub-intensity matrix.}
}
\value{
Laplace transform at \code{r}.
}
\description{
Computes the Laplace transform at \code{r} of a phase-type distribution with
 parameters \code{alpha} and \code{S}.
}
