% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phcdf}
\alias{phcdf}
\title{Phase-type cdf or tail}
\usage{
phcdf(x, pi, T, lower_tail = TRUE)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}
}
\value{
The cdf (tail) at \code{x}
}
\description{
Computes the cdf of phase-type distribution with parameters \code{pi} and \code{T} at \code{x}
}
\examples{
alpha <- c(0.5, 0.3, 0.2)
T <- matrix(c(c(-1,0,0),c(1,-2,0),c(0,1,-5)), nrow = 3, ncol = 3)
phcdf(0.5, alpha, T) 
phcdf(0.5, alpha, T, FALSE) 
}
