% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phdensity}
\alias{phdensity}
\title{Phase-type density}
\usage{
phdensity(x, pi, T)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}
}
\value{
The density at \code{x}
}
\description{
Computes the density of phase-type distribution with parameters \code{pi and} \code{T} at \code{x}
}
\examples{
alpha <- c(0.5, 0.3, 0.2)
T <- matrix(c(c(-1,0,0),c(1,-2,0),c(0,1,-5)), nrow = 3, ncol = 3)
t <- -T\%*\%rep(1, length(T[,1]))
n <- 10
phdensity(0.5, alpha, T, t) 
}
