% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phLaplace}
\alias{phLaplace}
\title{Laplace transform of a phase-type}
\usage{
phLaplace(s, pi, T)
}
\arguments{
\item{s}{real value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}
}
\value{
Laplace transform
}
\description{
Computes the Laplace transform at \code{s} of a phase-type distribution with parameters \code{pi} and \code{T}
}
\examples{
alpha <- c(0.5, 0.3, 0.2)
T <- matrix(c(c(-1,0,0),c(1,-2,0),c(0,1,-5)), nrow = 3, ncol = 3)
phLaplace(0.5, alpha, T) 
phLaplace(2.5, alpha, T) 
}
