% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phmoment}
\alias{phmoment}
\title{k moment of a phase-type}
\usage{
phmoment(k, pi, T)
}
\arguments{
\item{k}{Integer value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}
}
\value{
The k moment
}
\description{
Computes the k moment of phase-type distribution with parameters \code{pi} and \code{T}
}
