% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pi_null}
\alias{pi_null}
\title{Get pi0}
\usage{
pi_null(R, alpha)
}
\arguments{
\item{R}{reward vector}

\item{alpha}{initial distribution vector of the original MJP}
}
\value{
The initial distribution vector for states associated with null rewards
}
\description{
Get pi0
}
