% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pi_pos}
\alias{pi_pos}
\title{Get pi+}
\usage{
pi_pos(R, alpha)
}
\arguments{
\item{R}{reward vector}

\item{alpha}{initial distribution vector of the original MJP}
}
\value{
The initial distribution vector for states associated with positive rewards
}
\description{
Get pi+
}
