% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{plot,ph-method}
\alias{plot,ph-method}
\title{Plot Method for phase type distributions}
\usage{
\S4method{plot}{ph}(x, y = NULL)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{a dataset}
}
\description{
Plot Method for phase type distributions
}
\examples{

}
