% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{+,dph,dph-method}
\alias{+,dph,dph-method}
\title{Sum Method for discrete phase-type distributions}
\usage{
\S4method{+}{dph,dph}(e1, e2)
}
\arguments{
\item{e1}{An object of class \linkS4class{dph}.}

\item{e2}{An object of class \linkS4class{dph}.}
}
\value{
An object of class \linkS4class{dph}.
}
\description{
Sum Method for discrete phase-type distributions
}
\examples{
dph1 <- dph(structure = "general", dimension = 3)
dph2 <- dph(structure = "general", dimension = 5)
dph_sum <- dph1 + dph2
dph_sum
}
