% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{+,ph,ph-method}
\alias{+,ph,ph-method}
\title{Sum Method for phase type distributions}
\usage{
\S4method{+}{ph,ph}(e1, e2)
}
\arguments{
\item{e1}{an object of class \linkS4class{ph}.}

\item{e2}{an object of class \linkS4class{ph}.}
}
\description{
Sum Method for phase type distributions
}
