% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plus_states}
\alias{plus_states}
\title{Find which states have positive reward}
\usage{
plus_states(R)
}
\arguments{
\item{R}{reward vector}
}
\value{
A vector with the states (number) that are associated with positive rewards
}
\description{
Find which states have positive reward
}
