% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixt.R
\name{posmatsqrt}
\alias{posmatsqrt}
\title{posmatsqrt}
\usage{
posmatsqrt(A)
}
\arguments{
\item{A}{positive definite p x p real-valued matrix.}
}
\value{
a symmetric square root matrix for A. ie, \eqn{B = t(B)} and \eqn{B \%*\% B = A}.
}
\description{
posmatsqrt
}
\examples{
A = diag(5) + 1
B = posmatsqrt(A)
sum(abs(B - t(B)))
sum(abs(A - B \%*\% B))
}
