% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{q}
\alias{q}
\title{New Generic for the Quantile of Matrix Distributions}
\usage{
q(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{}
}
\value{
Quantile from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
