% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{q_null}
\alias{q_null}
\title{Obtain q0}
\usage{
q_null(R, Qtilda)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}
}
\value{
Creates the vector with transition probabilities from states associated with null rewards to the absorption state
}
\description{
Obtain q0
}
