% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Q_null_null}
\alias{Q_null_null}
\title{Obtain Q00}
\usage{
Q_null_null(R, Qtilda)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}
}
\value{
the matrix Q00 where we have transition probabilities from/to states associated with null rewards
}
\description{
Obtain Q00
}
