% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Q_null_pos}
\alias{Q_null_pos}
\title{Obtain Q0+}
\usage{
Q_null_pos(R, Qtilda)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}
}
\value{
the matrix Q0+ where we have transition probabilities from states associated with null rewards  to ones associated with positive rewards
}
\description{
Obtain Q0+
}
