% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{q_pos}
\alias{q_pos}
\title{Obtain q+}
\usage{
q_pos(R, Qtilda)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}
}
\value{
Creates the vector with transition probabilities from states associated with positive rewards to the absorption state
}
\description{
Obtain q+
}
