% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Q_pos_null}
\alias{Q_pos_null}
\title{Obtain Qtilda+0}
\usage{
Q_pos_null(R, Qtilda)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}
}
\value{
the matrix Qtilda+0 where we have transition probabilities from states associated with positive rewards to ones associated with null rewards
}
\description{
Obtain Qtilda+0
}
