% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Q_pos_pos}
\alias{Q_pos_pos}
\title{Obtain Q++}
\usage{
Q_pos_pos(R, Qtilda)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}
}
\value{
The matrix Q++ where we have transition probabilities from/to states associated with positive rewards
}
\description{
Obtain Q++
}
