% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{quan,ph-method}
\alias{quan,ph-method}
\title{Quantile method for phase-type distributions}
\usage{
\S4method{quan}{ph}(x, p)
}
\arguments{
\item{x}{An object of class \linkS4class{ph}.}

\item{p}{A vector of probabilities.}
}
\value{
A vector containing the quantile evaluations at the given locations.
}
\description{
Quantile method for phase-type distributions
}
\examples{
obj <- ph(structure = "general")
quan(obj, c(0.5, 0.9, 0.99))
}
