% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{quan}
\alias{quan}
\title{New generic for the quantile of matrix distributions}
\usage{
quan(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
Quantile from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
