% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
<<<<<<< HEAD
\docType{methods}
=======
>>>>>>> bfd4a4437aecc2d9b403fc560a9345bfda8bd07c
\name{r,iph-method}
\alias{r,iph-method}
\title{Simulation Method for inhomogeneous phase type distributions}
\usage{
\S4method{r}{iph}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{iph}.}

\item{n}{length of realization.}
}
\value{
A realization of inhomogeneous phase type data
}
\description{
Simulation Method for inhomogeneous phase type distributions
}
\examples{

}
