% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{r,mph-method}
\alias{r,mph-method}
\title{Simulation Method for multivariate phase type distributions}
\usage{
\S4method{r}{mph}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{mph}.}

\item{n}{length of realization.}
}
\value{
A realization of a phase type data
}
\description{
Simulation Method for multivariate phase type distributions
}
\examples{

}
