% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{r,ph-method}
\alias{r,ph-method}
\title{Simulation Method for phase type distributions}
\usage{
\S4method{r}{ph}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{n}{length of realization.}
}
\value{
A realization of a phase type data
}
\description{
Simulation Method for phase type distributions
}
\examples{

}
