% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{r}
\alias{r}
\title{New Generic for Simulating Matrix Distributions}
\usage{
r(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{}
}
\value{
A realization from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
