% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_phase_BivPH}
\alias{random_phase_BivPH}
\title{Random parameters of a bivariate phase-type}
\usage{
random_phase_BivPH(p1, p2, scale_factor = 1)
}
\description{
Random parameters of a bivariate phase-type
}
