% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_reward}
\alias{random_reward}
\title{Random reward matrix}
\usage{
random_reward(p, d)
}
\arguments{
\item{p}{Number of transient states in the sub-intensity matrix.}

\item{d}{Number of marginals.}
}
\value{
A random reward matrix.
}
\description{
Generates a random reward matrix for a multivariate phase-type distribution 
with p states and d marginals.
}
