% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_structure}
\alias{random_structure}
\title{Random structure of a phase-type}
\usage{
random_structure(p, structure = "General", scale_factor = 1)
}
\arguments{
\item{p}{Dimension of the phase-type}

\item{structure}{Type of structure: "General", "Hyperexponential", "GErlang", "Coxian" or "GCoxian"}

\item{scale_factor}{A factor that multiplies the sub-intensity matrix}
}
\value{
Random parameters \code{pi} and \code{T} of a phase-type
}
\description{
Generates random parameters \code{pi} and \code{T} of a phase-type distribution of dimension \code{p} with chosen structure
}
\examples{
random_structure(3) 
random_structure(5, "Hyperexponential") 
}
