% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{random_structure_bivph}
\alias{random_structure_bivph}
\title{Random structure of a bivariate phase-type}
\usage{
random_structure_bivph(p1, p2, scale_factor = 1)
}
\arguments{
\item{p1}{Dimension of the first block.}

\item{p2}{Dimension of the second block.}

\item{scale_factor}{A factor that multiplies the sub-intensity matrix.}
}
\value{
Random parameters  \code{alpha}, \code{S11}, \code{S12}, and \code{S22}
 of a bivariate phase-type.
}
\description{
Generates random parameters \code{alpha}, \code{S11}, \code{S12}, and \code{S22}
of a bivariate phase-type distribution of dimension \code{p  = p1 + p2}.
}
