% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcsph}
\alias{rcsph}
\title{Simulate common-shock phase-type}
\usage{
rcsph(n, alpha, S, P, Q1, Q2, a1, a2)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Vector of initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{P}{Matrix.}

\item{Q1}{Sub-intensity matrix.}

\item{Q2}{Sub-intensity matrix.}

\item{a1}{Multiplicative constant for margin 1.}

\item{a2}{Multiplicative constant for margin 1.}
}
\value{
Simulated sample.
}
\description{
Generates a sample of size \code{n} from a phase-type distribution with
parameters \code{alpha}, \code{S}, \code{P}, \code{Q1}, \code{Q2},
\code{a1}, and \code{a2}
}
