% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rdphasetype}
\alias{rdphasetype}
\title{Simulate discrete phase-type}
\usage{
rdphasetype(n, alpha, S)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Vector of initial probabilities.}

\item{S}{Sub-transition matrix.}
}
\value{
Simulated sample.
}
\description{
Generates a sample of size \code{n} from a discrete phase-type distribution with
parameters \code{alpha} and \code{S}.
}
