% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b.iph_regression.R
\name{reg,ph-method}
\alias{reg,ph-method}
\title{Regression Method for ph Class}
\usage{
\S4method{reg}{ph}(
  x,
  y,
  weight = numeric(0),
  rcen = numeric(0),
  rcenweight = numeric(0),
  X = numeric(0),
  B0 = numeric(0),
  stepsEM = 1000
)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{vector or data.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Regression Method for ph Class
}
\examples{

}
