% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.iph_double_regression.R
\name{reg2,iph-method}
\alias{reg2,iph-method}
\title{Regression Method for ph Class}
\usage{
\S4method{reg2}{iph}(
  x,
  y,
  weight = numeric(0),
  rcen = numeric(0),
  rcenweight = numeric(0),
  X = numeric(0),
  Z = numeric(0),
  stepsEM = 1000,
  B0 = numeric(0),
  C0 = numeric(0)
)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{vector or data.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Regression Method for ph Class
}
\examples{

}
