% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{reg2}
\alias{reg2}
\title{New Generic for Double Regression with Matrix Distributions}
\usage{
reg2(x, y, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{y}{a vector of data.}

\item{...}{}
}
\value{
An object of the fitted model class.
}
\description{
Methods are available for objects of class \linkS4class{iph}
}
