% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b.iph_regression.R
\name{reg_newton,ph-method}
\alias{reg_newton,ph-method}
\title{Regression Method for ph Class using Newton-Rhapson}
\usage{
\S4method{reg_newton}{ph}(
  x,
  y,
  weight = numeric(0),
  X = numeric(0),
  B0 = numeric(0),
  method = "UNI",
  rkstep = NA,
  uni_epsilon = NA,
  h = 1,
  stepsEM = 1000,
  stepsNR = 5,
  every = 10
)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{vector or data.}

\item{weight}{vector of weights.}

\item{X}{model matrix (no intercept needed).}

\item{B0}{initial regression coefficients (optional).}

\item{method}{method for the EM steps}

\item{rkstep}{Runge-Kutta step size (optional)}

\item{uni_epsilon}{epsilon parameter for uniformization method}

\item{h}{parameter for Newton-Raphson.}

\item{stepsEM}{number of EM steps to be performed.}

\item{every}{number of iterations between likelihood display updates.}
}
\value{
An object of class \linkS4class{sph}.
}
\description{
Regression Method for ph Class using Newton-Rhapson
}
