% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{revers_data_trans}
\alias{revers_data_trans}
\title{Applies the inverse of the GEV transformation but giving back the resulting
vector in reverse order}
\usage{
revers_data_trans(obs, weights, beta)
}
\arguments{
\item{obs}{The observations.}

\item{weights}{Weights of the observations.}

\item{beta}{Parameters of the GEV.}
}
\description{
Used for EM step in RK.
}
