% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reversTransformData}
\alias{reversTransformData}
\title{Applies the inverse of the GEV but giving back the vector in reverse order}
\usage{
reversTransformData(observations, weights, beta)
}
\description{
Applies the inverse of the GEV but giving back the vector in reverse order
}
