% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rew_sanity_check}
\alias{rew_sanity_check}
\title{Transform a reward matrix with very small rewards to avoid numerical problems}
\usage{
rew_sanity_check(R, tol)
}
\arguments{
\item{R}{Reward matrix}

\item{tol}{Lower bound considered for a reward}
}
\value{
A reward matrix that does not cause issues with uniformization
}
\description{
Transform a reward matrix with very small rewards to avoid numerical problems
}
