% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rimph}
\alias{rimph}
\title{Random Inhomogeneous MPH*}
\usage{
rimph(n, dist_type, pi, T, R, beta)
}
\arguments{
\item{n}{Sample size}

\item{dist_type}{Type of distribution: "Weibull" "Pareto"}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{parameters of the transformations}
}
\value{
The simulated sample
}
\description{
Generates a sample of size \code{n} from an Inhomogeneous MPH* distribution with parameters \code{pi}, \code{T} and \code{R}
}
\examples{
alpha <- c(0.5, 0.3, 0.2)
T <- matrix(c(c(-1,0,0),c(1,-2,0),c(0,1,-5)), nrow = 3, ncol = 3)
R <- matrix(c(c(1,0,0.8),c(0,1,0.2)), nrow = 3, ncol = 2)
beta <- c(0.4, 0.7)
n <- 10
rimph(n, "Weibull", alpha, T, R, beta) 
}
