% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{riph}
\alias{riph}
\title{Random inhomogeneous phase-type}
\usage{
riph(n, dist_type, pi, T, beta)
}
\arguments{
\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{Parameter of the transformation}
}
\value{
The simulated sample
}
\description{
Generates a sample of size \code{n} from an inhomogeneous phase-type distribution with parameters \code{pi}, \code{T} and \code{beta}
}
\examples{
alpha <- c(0.5, 0.3, 0.2)
T <- matrix(c(c(-1,0,0),c(1,-2,0),c(0,1,-5)), nrow = 3, ncol = 3)
beta <- 0.5
n <- 10
riph(n, "Pareto", alpha, T, beta) 
}
