% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmatrixgev}
\alias{rmatrixgev}
\title{Random matrix GEV}
\usage{
rmatrixgev(n, alpha, S, mu, sigma, xi = 0)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{xi}{Shape parameter: Default 0 which corresponds to the Gumbel case.}
}
\value{
The simulated sample.
}
\description{
Generates a sample of size \code{n} from an inhomogeneous phase-type 
distribution with parameters \code{alpha}, \code{S} and \code{beta}.
}
