% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixnorm.R
\name{rmatrixnorm}
\alias{rmatrixnorm}
\title{rmatrixnorm}
\usage{
rmatrixnorm(n, mean, L = diag(dim(mean)[1]), R = diag(dim(mean)[2]), U = L
  \%*\% t(L), V = t(R) \%*\% R, list = FALSE, array = NULL)
}
\arguments{
\item{n}{number of observations to generate - must be a positive integer.}

\item{mean}{$p \times q$ matrix of means}

\item{L}{$p \times p$ matrix specifying relations among the rows. By default, an identity matrix.}

\item{R}{$q \times q$ matrix specifying relations among the columns. By default, an identity matrix.}

\item{U}{$LL^T$ - $p \times p$ positive definite variance-covariance matrix for rows, computed from $L$ if not specified.}

\item{V}{$R^TR$ - $q \times q$ positive definite variance-covariance matrix for columns, computed from $R$ if not specified.}

\item{list}{Defaults to \code{FALSE}. If this is \code{TRUE}, then the output will be a list of matrices.}

\item{array}{If $n = 1$ and this is not specified and \code{list} is \code{FALSE}, the function will return a matrix containing the one observation. If $n > 1$, should be the opposite of \code{list}. If \code{list} is \code{TRUE}, this will be ignored.}
}
\value{
This returns either a list of $n$ $p \times q$ matrices or a $n \times p \times q$ array.
}
\description{
rmatrixnorm
}
