% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixt.R
\name{rmatrixt}
\alias{rmatrixt}
\title{Title}
\usage{
rmatrixt(n, mean, df, L = diag(dim(mean)[1]), R = diag(dim(mean)[2]),
  U = L \%*\% t(L), V = t(R) \%*\% R, list = FALSE, array = NULL)
}
\arguments{
\item{n}{number of observations to generate}

\item{mean}{\eqn{p X q} This is really a "shift" rather than a mean as this
is a central T, though the expected value will be equal to this if \eqn{df > 2}}

\item{df}{Degrees of freedom parameter}

\item{L}{\eqn{p X p}  matrix specifying relations among the rows. By default, an identity matrix.}

\item{R}{\eqn{q X q}  matrix specifying relations among the columns. By default, an identity matrix.}

\item{U}{\eqn{LL^T}  - \eqn{p X p}  positive definite matrix for rows, computed
from \eqn{L} if not specified. Note this is not the variance.}

\item{V}{\eqn{R^T R}  - \eqn{q X q}  positive definite matrix for columns,
computed from \eqn{R}  if not specified.  Note this is not the variance.}

\item{list}{Defaults to \code{FALSE} . If this is \code{TRUE} , then the output will be a list of matrices.}

\item{array}{If \eqn{n = 1}  and this is not specified and \code{list}  is \code{FALSE} , the function will return a matrix containing the one observation. If \eqn{n > 1} , should be the opposite of \code{list} . If \code{list}  is \code{TRUE} , this will be ignored.}
}
\value{
This returns either a list of \eqn{n}  \eqn{p X q}  matrices or a \eqn{n X p X q}  array.
}
\description{
Title
}
\examples{
set.seed(20180202)
rmatrixt(n=1,mean=matrix(c(100,0,-100,0,25,-1000),nrow=2),df=10,
   L=matrix(c(2,1,0,.1),nrow=2),list=FALSE)
set.seed(20180202)
A = rmatrixt(n=10,mean=matrix(c(100,0,-100,0,25,-1000),nrow=2),df=10,
   L=matrix(c(2,1,0,.1),nrow=2),list=TRUE)
A[[1]]
set.seed(20180202)
B = rmatrixt(n=10,mean=matrix(c(100,0,-100,0,25,-1000),nrow=2),df=10,
   L=matrix(c(2,1,0,.1),nrow=2),list=FALSE)
B[1 , , ]
summary(rt(n=100,df=10))
summary(rmatrixt(n=100,matrix(0),df=10))

}
