% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rMDPHstar}
\alias{rMDPHstar}
\title{Simulate MDPH*}
\usage{
rMDPHstar(n, alpha, S, R)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Vector of initial probabilities.}

\item{S}{Sub-transition matrix.}

\item{R}{Reward matrix.}
}
\value{
Simulated sample.
}
\description{
Generates a sample of size \code{n} from a MDPH* distribution with
parameters \code{alpha}, \code{S}, and \code{R}.
}
