% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rMIPHstar}
\alias{rMIPHstar}
\title{Simulate a MIPH* random vector}
\usage{
rMIPHstar(n, alpha, S, R, gfun, gfun_par)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{R}{Reward matrix.}

\item{gfun}{Vector with transformations names.}

\item{gfun_par}{List with transformations parameters.}
}
\value{
The simulated sample.
}
\description{
Generates a sample of size \code{n} from a MIPH* distribution with parameters
\code{alpha}, \code{S} and \code{R}.
}
