% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmph}
\alias{rmph}
\title{Random MPH*}
\usage{
rmph(n, pi, T, R)
}
\arguments{
\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}
}
\value{
The simulated sample
}
\description{
Generates a sample of size \code{n} from a MPH* distribution with parameters \code{pi}, \code{T} and \code{R}
}
\examples{
alpha <- c(0.5, 0.3, 0.2)
T <- matrix(c(c(-1,0,0),c(1,-2,0),c(0,1,-5)), nrow = 3, ncol = 3)
R <- matrix(c(c(1,0,0.8),c(0,1,0.2)), nrow = 3, ncol = 2)
n <- 10
rmph(n, alpha, T, R) 
}
