% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rMPHstar}
\alias{rMPHstar}
\title{Random multivariate phase-type via rewards}
\usage{
rMPHstar(n, alpha, S, R)
}
\arguments{
\item{n}{Sample size for each marginal.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{R}{Reward matrix.}
}
\value{
The simulated samples, each column corresponds to a marginal.
}
\description{
Generates samples of size \code{n} from a multivariate phase-type distribution
 with parameters \code{alpha} and \code{S}.
}
