% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runge_kutta}
\alias{runge_kutta}
\title{Runge Kutta for the calculation of the a,b and c vectors in a EM step}
\usage{
runge_kutta(avector, bvector, cmatrix, dt, h, T, t)
}
\description{
I may need to change the type of avector and bvector, depending on how I call them in the EM step
}
