% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runge_kutta_new}
\alias{runge_kutta_new}
\title{Runge Kutta for the calculation of the a,b and c vectors in a EM step}
\usage{
runge_kutta_new(avector, bvector, cmatrix, dt, h, S, s)
}
\arguments{
\item{avector}{The a vector.}

\item{bvector}{The b vector.}

\item{cmatrix}{The c matrix.}

\item{dt}{The increment.}

\item{h}{Step-length.}

\item{S}{Sub-intensity.}

\item{s}{Exit rates.}
}
\description{
Performs the RK of forth order
}
