% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{safe_cov}
\alias{safe_cov}
\title{Calculate Covariance Matrix Safely}
\usage{
safe_cov(hess)
}
\arguments{
\item{hess}{a Hessian matrix from a model fit.}
}
\value{
Fisher information (estimated covariance matrix)
}
\description{
Calculate Covariance Matrix Safely
}
