% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{secondEMstep}
\alias{secondEMstep}
\title{Second EM in the algorithm of bruer}
\usage{
secondEMstep(observations, weight, censored, rcweight, pi, T, R)
}
\description{
Second EM in the algorithm of bruer
}
