% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\name{show,bivdph-method}
\alias{show,bivdph-method}
\title{Show method for bivariate discrete phase-type distributions}
\usage{
\S4method{show}{bivdph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{bivdph}.}
}
\description{
Show method for bivariate discrete phase-type distributions
}
