% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4c.biviph.R
\name{show,biviph-method}
\alias{show,biviph-method}
\title{Show method for bivariate inhomogeneous phase-type distributions}
\usage{
\S4method{show}{biviph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{biviph}.}
}
\description{
Show method for bivariate inhomogeneous phase-type distributions
}
