% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.bivph.R
\name{show,bivph-method}
\alias{show,bivph-method}
\title{Show method for bivariate phase-type distributions}
\usage{
\S4method{show}{bivph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{bivph}.}
}
\description{
Show method for bivariate phase-type distributions
}
