% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.bph.R
\name{show,bph-method}
\alias{show,bph-method}
\title{Show Method for bivariate phase type distributions}
\usage{
\S4method{show}{bph}(object)
}
\arguments{
\item{x}{an object of class \linkS4class{bph}.}
}
\description{
Show Method for bivariate phase type distributions
}
\examples{

}
