% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{show,dph-method}
\alias{show,dph-method}
\title{Show Method for discrete phase-type distributions}
\usage{
\S4method{show}{dph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{dph}.}
}
\description{
Show Method for discrete phase-type distributions
}
