% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.ibph.R
\name{show,ibph-method}
\alias{show,ibph-method}
\title{Show Method for inhomogeneous bivariate phase type distributions}
\usage{
\S4method{show}{ibph}(object)
}
\arguments{
\item{x}{an object of class \linkS4class{ibph}.}
}
\description{
Show Method for inhomogeneous bivariate phase type distributions
}
\examples{

}
