% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{show,mdph-method}
\alias{show,mdph-method}
\title{Show Method for multivariate discrete phase-type distributions}
\usage{
\S4method{show}{mdph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{mdph}.}
}
\description{
Show Method for multivariate discrete phase-type distributions
}
