% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b.miph.R
\name{show,miph-method}
\alias{show,miph-method}
\title{Show Method for multivariate inhomogeneous phase-type distributions}
\usage{
\S4method{show}{miph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{miph}.}
}
\description{
Show Method for multivariate inhomogeneous phase-type distributions
}
