% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{show,MPHstar-method}
\alias{show,MPHstar-method}
\title{Show method for multivariate phase-type distributions}
\usage{
\S4method{show}{MPHstar}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{MPHstar}.}
}
\description{
Show method for multivariate phase-type distributions
}
