% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{show,ph-method}
\alias{show,ph-method}
\title{Show method for phase-type distributions}
\usage{
\S4method{show}{ph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{ph}.}
}
\description{
Show method for phase-type distributions
}
