% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.sph.R
\name{show,sph-method}
\alias{show,sph-method}
\title{Show Method for survival phase-type objects}
\usage{
\S4method{show}{sph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{sph}.}
}
\description{
Show Method for survival phase-type objects
}
