% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\name{sim,bivdph-method}
\alias{sim,bivdph-method}
\title{Simulation method for bivariate discrete phase-type distributions}
\usage{
\S4method{sim}{bivdph}(x, n = 1000)
}
\arguments{
\item{x}{An object of class \linkS4class{bivdph}.}

\item{n}{An integer of length of realization.}
}
\value{
A realization of independent and identically distributed bivariate
discrete phase-type vector.
}
\description{
Simulation method for bivariate discrete phase-type distributions
}
\examples{
obj <- bivdph(dimensions = c(3, 3))
sim(obj, n = 100)
}
