% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4c.biviph.R
\name{sim,biviph-method}
\alias{sim,biviph-method}
\title{Simulation method for bivariate inhomogeneous phase-type distributions}
\usage{
\S4method{sim}{biviph}(x, n = 1000)
}
\arguments{
\item{x}{An object of class \linkS4class{biviph}.}

\item{n}{An integer of length of realization.}
}
\value{
A realization of independent and identically distributed bivariate
inhomogeneous phase-type vector.
}
\description{
Simulation method for bivariate inhomogeneous phase-type distributions
}
\examples{
under_bivph <- bivph(dimensions = c(3, 3))
obj <- biviph(under_bivph, gfun = c("weibull", "pareto"), gfun_pars = list(c(2), c(3)))
sim(obj, n = 100)
}
