% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.ibph.R
\name{sim,ibph-method}
\alias{sim,ibph-method}
\title{Simulation Method for inhomogeneous bivariate phase type distributions}
\usage{
\S4method{sim}{ibph}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{ibph}.}

\item{n}{length of realization.}
}
\value{
A realization of inhomogeneous phase type data
}
\description{
Simulation Method for inhomogeneous bivariate phase type distributions
}
\examples{

}
