% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{sim,iph-method}
\alias{sim,iph-method}
\title{Simulation method for inhomogeneous phase-type distributions}
\usage{
\S4method{sim}{iph}(x, n = 1000)
}
\arguments{
\item{x}{An object of class \linkS4class{iph}.}

\item{n}{An integer of length of realization.}
}
\value{
A realization of independent and identically distributed inhomogeneous
phase-type variables.
}
\description{
Simulation method for inhomogeneous phase-type distributions
}
\examples{
obj <- iph(ph(structure = "general"), gfun = "lognormal", gfun_pars = 2)
sim(obj, n = 100)
}
