% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{sim,mdph-method}
\alias{sim,mdph-method}
\title{Simulation method for multivariate discrete phase-type distributions}
\usage{
\S4method{sim}{mdph}(x, n = 1000, equal_marginals = 0)
}
\arguments{
\item{x}{An object of class \linkS4class{mdph}.}

\item{n}{Length of realization.}

\item{equal_marginals}{Non-negative integer. If positive, it specifies
the number of marginals to simulate from, all from the first matrix.}
}
\value{
A realization of a multivariate discrete phase-type distribution.
}
\description{
Simulation method for multivariate discrete phase-type distributions
}
\examples{
obj <- mdph(structure = c("general", "general"))
sim(obj, 100)
}
