% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{sim,MPHstar-method}
\alias{sim,MPHstar-method}
\title{Simulation method for multivariate phase-type distributions}
\usage{
\S4method{sim}{MPHstar}(x, n = 1000)
}
\arguments{
\item{x}{An object of class \linkS4class{MPHstar}.}

\item{n}{Desired sample size for each marginal.}
}
\value{
A matrix of sample data for each marginal.
}
\description{
Simulation method for multivariate phase-type distributions
}
\examples{
obj <- MPHstar(structure = "general")
sim(obj, 100)
}
