% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{sim,ph-method}
\alias{sim,ph-method}
\title{Simulation Method for phase type distributions}
\usage{
\S4method{sim}{ph}(x, n = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{n}{length of realization.}
}
\value{
A realization of a phase type data
}
\description{
Simulation Method for phase type distributions
}
\examples{

}
