% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.sph.R
\docType{class}
\name{sph-class}
\alias{sph-class}
\title{Survival Analysis for Phase Type distributions}
\value{

}
\description{
Class of objects for inhomogeneous phase type distributions
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the phase type distribution}

\item{\code{pars}}{a list comprising of the parameters.}
}}

