% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_ph}
\alias{sum_ph}
\title{Computes the initial distribution and sub-intensity of the sum of PH}
\usage{
sum_ph(alpha1, S1, alpha2, S2)
}
\arguments{
\item{alpha1}{Initial distribution.}

\item{S1}{Sub-intensity.}

\item{alpha2}{Initial distribution.}

\item{S2}{Sub-intensity.}
}
\description{
Computes the initial distribution and sub-intensity of the sum of PH
}
