% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumPH}
\alias{sumPH}
\title{Computes the initial distriubtion and sub-intensity of the sum of PH}
\usage{
sumPH(pi1, T1, pi2, T2)
}
\arguments{
\item{pi1}{initial distribution}

\item{T1}{sub-intensity}

\item{pi2}{initial distribution}

\item{T2}{sub-intensity}
}
\description{
Computes the initial distriubtion and sub-intensity of the sum of PH
}
