% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixnorm.R
\name{toepmatrix}
\alias{toepmatrix}
\title{toepmatrix}
\usage{
toepmatrix(n, rho)
}
\arguments{
\item{n}{number of columns/rows}

\item{rho}{correlation parameter}
}
\value{
Toeplitz $n x n$ matrix with 1 on the diagonal and $rho^k$ on the other diagonals, where $k$ is distance from the main diagonal. Used internally but it is useful for generating your own random matrices.
}
\description{
toepmatrix
}
\examples{
rho = .9
n = 6
toepmatrix(n,rho)
}
\keyword{internal}
