% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{transf_via_rew}
\alias{transf_via_rew}
\title{Performs the TVR}
\usage{
transf_via_rew(R, Qtilda, alpha, S)
}
\arguments{
\item{R}{reward vector}

\item{Qtilda}{transition-matrix of the Embedded MC}

\item{alpha}{initial distribution vector}

\item{S}{original sub-intensity matrix}
}
\value{
A list of transformed PH distributions
}
\description{
Performs the TVR
}
