% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{TVR,dph-method}
\alias{TVR,dph-method}
\title{TVR Method for dph Class}
\usage{
\S4method{TVR}{dph}(x, rew)
}
\arguments{
\item{x}{An object of class \linkS4class{dph}.}

\item{rew}{A vector of rewards.}
}
\value{
An object of the of class \linkS4class{dph}.
}
\description{
TVR Method for dph Class
}
\examples{
obj <- dph(structure = "general")
TVR(obj, c(1, 0, 1))
}
