% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{TVR,MPHstar-method}
\alias{TVR,MPHstar-method}
\title{Transformation Via Rewards method for multivariate phase type distributions}
\usage{
\S4method{TVR}{MPHstar}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{MPHstar}.}
}
\value{
A list with marginal distributions, obtained by transformation via rewards.
}
\description{
Transformation Via Rewards method for multivariate phase type distributions
}
