% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{TVR,ph-method}
\alias{TVR,ph-method}
\title{TVR method for ph class}
\usage{
\S4method{TVR}{ph}(x, rew)
}
\arguments{
\item{x}{An object of class \linkS4class{ph}.}

\item{rew}{A vector of rewards.}
}
\value{
An object of the of class \linkS4class{ph}.
}
\description{
TVR method for ph class
}
\examples{
obj <- ph(structure = "general")
TVR(obj, c(1, 2, 3))
}
