% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tvr_dph}
\alias{tvr_dph}
\title{Performs TVR for discrete phase-type distributions}
\usage{
tvr_dph(alpha, S, R)
}
\arguments{
\item{alpha}{Initial distribution vector.}

\item{S}{Sub-intensity matrix.}

\item{R}{Reward vector.}
}
\value{
A list of PH parameters.
}
\description{
Performs TVR for discrete phase-type distributions
}
