% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.bivdph.R
\name{var,bivdph-method}
\alias{var,bivdph-method}
\title{Var Method for bivdph class}
\usage{
\S4method{var}{bivdph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{bivdph}.}
}
\value{
The covariance matrix of the bivariate discrete phase-type distribution.
}
\description{
Var Method for bivdph class
}
\examples{
obj <- bivdph(dimensions = c(3, 3))
var(obj)
}
