% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.bivph.R
\name{var,bivph-method}
\alias{var,bivph-method}
\title{Var method for bivph class}
\usage{
\S4method{var}{bivph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{bivph}.}
}
\value{
The covariance matrix of the bivariate phase-type distribution.
}
\description{
Var method for bivph class
}
\examples{
obj <- bivph(dimensions = c(3, 3))
var(obj)
}
