% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{var,mdph-method}
\alias{var,mdph-method}
\title{Var method for multivariate discrete phase-type distributions}
\usage{
\S4method{var}{mdph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{mdph}.}
}
\value{
The covariance matrix of the multivariate discrete phase-type distribution.
}
\description{
Var method for multivariate discrete phase-type distributions
}
\examples{
obj <- mdph(structure = c("general", "general"))
var(obj)
}
