% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{var,mph-method}
\alias{var,mph-method}
\title{Var Method for multivariate phase-type distributions}
\usage{
\S4method{var}{mph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}
}
\value{
The covariance matrix of the multivariate phase-type distribution.
}
\description{
Var Method for multivariate phase-type distributions
}
\examples{
obj <- mph(structure = c("general", "general"))
var(obj)
}
