% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{var,ph-method}
\alias{var,ph-method}
\title{Var Method for phase-type distributions}
\usage{
\S4method{var}{ph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{ph}.}
}
\value{
The variance of the \linkS4class{ph} (or undelying \linkS4class{ph}) object.
}
\description{
Var Method for phase-type distributions
}
\examples{
set.seed(123)
ph1 <- ph(structure = "general", dimension = 3)
var(ph1)
}
