% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_of_matrices}
\alias{vector_of_matrices}
\title{Computes elements S^n / n! until the value size}
\usage{
vector_of_matrices(theVector, S, a, sizevect)
}
\arguments{
\item{theVector}{A vector.}

\item{S}{Sub-untensity matrix.}

\item{a}{A number.}

\item{sizevect}{Size of vector.}
}
\description{
Computes elements S^n / n! until the value size
}
