% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_of_matrices_2}
\alias{vector_of_matrices_2}
\title{Computes elements S^n / n! until the value size}
\usage{
vector_of_matrices_2(theVector, S, vect_size)
}
\arguments{
\item{theVector}{A vector.}

\item{S}{Sub-intensity matrix.}

\item{vect_size}{Size of vector.}
}
\description{
Computes elements S^n / n! until the value size
}
