% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_of_powers}
\alias{vector_of_powers}
\title{Computes elements A^n until the given size}
\usage{
vector_of_powers(A, vect_size)
}
\arguments{
\item{A}{A matrix.}

\item{vect_size}{Size of vector.}
}
\description{
Computes elements A^n until the given size
}
