% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectorOfMatrices_arma}
\alias{vectorOfMatrices_arma}
\title{Computes elements S^n / n! until the value size}
\usage{
vectorOfMatrices_arma(theVector, S, a, sizevect)
}
\arguments{
\item{theVector}{a vector}

\item{S}{sub-untensity matrix}

\item{a}{a number}

\item{sizevect}{size of vector}
}
\description{
Computes elements S^n / n! until the value size
}
