% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vectorOfMatrices_arma2}
\alias{vectorOfMatrices_arma2}
\title{Computes elements S^n / n! until the value size}
\usage{
vectorOfMatrices_arma2(theVector, S, sizevect)
}
\arguments{
\item{theVector}{a vector}

\item{S}{sub-untensity matrix}

\item{sizevect}{size of vector}
}
\description{
Computes elements S^n / n! until the value size
}
