% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{a_rungekutta}
\alias{a_rungekutta}
\title{Runge Kutta for the calculation of the a vectors in a EM step - Can be used for the loglikelihood}
\usage{
a_rungekutta(avector, dt, h, T)
}
\description{
I may need to change the type of avector
}
