% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{a_rungekutta}
\alias{a_rungekutta}
\title{Runge-Kutta for the calculation of the a vector in a EM step}
\usage{
a_rungekutta(avector, dt, h, S)
}
\arguments{
\item{avector}{The a vector.}

\item{dt}{Increment.}

\item{h}{Step-length.}

\item{S}{Sub-intensity matrix.}
}
\description{
Runge-Kutta for the calculation of the a vector in a EM step
}
