% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{cdf,iph-method}
\alias{cdf,iph-method}
\title{Distribution method for inhomogeneous phase-type distributions}
\usage{
\S4method{cdf}{iph}(x, q, lower.tail = TRUE)
}
\arguments{
\item{x}{An object of class \linkS4class{iph}.}

\item{q}{A vector of locations.}

\item{lower.tail}{Logical parameter specifying whether lower tail (CDF) or
upper tail is computed.}
}
\value{
A vector containing the CDF evaluations at the given locations.
}
\description{
Distribution method for inhomogeneous phase-type distributions
}
\examples{
obj <- iph(ph(structure = "general"), gfun = "weibull", gfun_pars = 2)
cdf(obj, c(1, 2, 3))
}
