% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{cdf,ph-method}
\alias{cdf,ph-method}
\title{Distribution Method for phase type distributions}
\usage{
\S4method{cdf}{ph}(x, q = seq(0, quan(x, 0.95)$quantile, length.out = 10), lower.tail = TRUE)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{q}{a vector of locations.}

\item{lower.tail}{logical parameter specifying whether lower tail (cdf) or upper tail is computed.}
}
\value{
A list containing the locations and corresponding CDF evaluations.
}
\description{
Distribution Method for phase type distributions
}
\examples{
obj <- ph(structure = "general")
cdf(obj, c(1, 2, 3))
}
