% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{coef,iph-method}
\alias{coef,iph-method}
\title{Coef Method for iph Class}
\usage{
\S4method{coef}{iph}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{iph}.}
}
\value{
parameters of iph model.
}
\description{
Coef Method for iph Class
}
\examples{
obj <- iph(ph(structure = "general", dimension = 2), gfun = "lognormal", gfun_pars = 2)
coef(obj)
}
