% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{coef,ph-method}
\alias{coef,ph-method}
\title{Coef method for ph class}
\usage{
\S4method{coef}{ph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{ph}.}
}
\value{
Parameters of ph model.
}
\description{
Coef method for ph class
}
\examples{
obj <- ph(structure = "general")
coef(obj)
}
