% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{dens,iph-method}
\alias{dens,iph-method}
\title{Density Method for inhomogeneous phase type distributions}
\usage{
\S4method{dens}{iph}(x, y = seq(0, quan(x, 0.95)$quantile, length.out = 10))
}
\arguments{
\item{x}{an object of class \linkS4class{iph}.}

\item{y}{a vector of locations.}
}
\value{
A list containing the locations and corresponding density evaluations.
}
\description{
Density Method for inhomogeneous phase type distributions
}
\examples{
obj <- iph(ph(structure = "general"), gfun = "weibull", gfun_pars = 2)
dens(obj, c(1, 2, 3))
}
