% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{dens,iph-method}
\alias{dens,iph-method}
\title{Density method for inhomogeneous phase-type distributions}
\usage{
\S4method{dens}{iph}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{iph}.}

\item{y}{A vector of locations.}
}
\value{
A vector containing the density evaluations at the given locations.
}
\description{
Density method for inhomogeneous phase-type distributions
}
\examples{
obj <- iph(ph(structure = "general"), gfun = "weibull", gfun_pars = 2)
dens(obj, c(1, 2, 3))
}
