% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b.miph.R
\name{dens,miph-method}
\alias{dens,miph-method}
\title{Density Method for multivariate inhomogeneous phase-type distributions}
\usage{
\S4method{dens}{miph}(x, y, delta = NULL)
}
\arguments{
\item{x}{An object of class \linkS4class{miph}.}

\item{y}{A matrix of observations.}

\item{delta}{Matrix with right-censoring indicators (1 uncensored, 0 right censored).}
}
\value{
A list containing the locations and corresponding density evaluations.
}
\description{
Density Method for multivariate inhomogeneous phase-type distributions
}
\examples{
under_mph <- mph(structure = c("general", "general"))
obj <- miph(under_mph, gfun = c("weibull", "pareto"), gfun_pars = list(c(2), c(3)))
dens(obj, c(1, 2))
}
