% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{dens,mph-method}
\alias{dens,mph-method}
\title{Density Method for multivariate phase-type distributions}
\usage{
\S4method{dens}{mph}(x, y, delta = NULL)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{y}{A matrix of observations.}

\item{delta}{Matrix with right-censoring indicators (1 uncensored, 0 right censored).}
}
\value{
A list containing the locations and corresponding density evaluations.
}
\description{
Density Method for multivariate phase-type distributions
}
