% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{dens,ph-method}
\alias{dens,ph-method}
\title{Density method for phase-type distributions}
\usage{
\S4method{dens}{ph}(x, y)
}
\arguments{
\item{x}{An object of class \linkS4class{ph}.}

\item{y}{A vector of locations.}
}
\value{
A vector containing the density evaluations at the given locations.
}
\description{
Density method for phase-type distributions
}
\examples{
obj <- ph(structure = "general")
dens(obj, c(1, 2, 3))
}
