% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derivativeMatrixWeibull}
\alias{derivativeMatrixWeibull}
\title{Derivative of matrix Weibull}
\usage{
derivativeMatrixWeibull(h, obs, weight, rcens, rcweight, pi, T, beta)
}
\description{
Derivative of matrix Weibull
}
