% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{derivativeMatrixweibull}
\alias{derivativeMatrixweibull}
\title{Derivative of matrix Weibull}
\usage{
derivativeMatrixweibull(h, obs, weight, rcens, rcweight, alpha, S, beta)
}
\arguments{
\item{h}{step-length}

\item{obs}{the observations}

\item{weight}{weight of the observations}

\item{rcens}{censored observations}

\item{rcweight}{weight of the censored observations}

\item{alpha}{initial probabilities}

\item{S}{sub-intensity}

\item{beta}{parameter of transformation}
}
\description{
Can be used to increase performance
}
