% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diagonal_vector}
\alias{diagonal_vector}
\title{Creates a matrix with the given vector in the diagonal}
\usage{
diagonal_vector(vec)
}
\arguments{
\item{vec}{a vector}
}
\description{
Creates a matrix with the given vector in the diagonal
}
