% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{dph}
\alias{dph}
\title{Constructor Function for discrete phase-type distributions}
\usage{
dph(alpha = NULL, S = NULL, structure = NULL, dimension = 3)
}
\arguments{
\item{alpha}{A probability vector.}

\item{S}{A sub-transition matrix.}

\item{structure}{A valid dph structure ("general", "coxian", "hyperexponential", "gcoxian", "gerlang").}

\item{dimension}{The dimension of the dph structure (if structure is provided).}
}
\value{
An object of class \linkS4class{dph}.
}
\description{
Constructor Function for discrete phase-type distributions
}
\examples{
dph(structure = "general", dim = 5)
dph(alpha = c(0.5, 0.5), S = matrix(c(0.1, 0.5, 0.5, 0.2), 2, 2))
}
