% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_PADE}
\alias{EMstep_PADE}
\title{EM using Matlab algorithm for matrix exponential in combination with Armadillo}
\usage{
EMstep_PADE(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Nuisance parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity.}

\item{obs}{The observations.}

\item{weight}{The weights for the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{The weights for the censored observations.}
}
\description{
EM using Matlab algorithm for matrix exponential in combination with Armadillo
}
