% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_PADE}
\alias{EMstep_PADE}
\title{EM for phase-type distributions using Pade approximation for matrix exponential}
\usage{
EMstep_PADE(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Nuisance parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{obs}{The observations.}

\item{weight}{The weights for the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{The weights for the censored observations.}
}
\description{
EM for phase-type distributions using Pade approximation for matrix exponential
}
