% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_PADE}
\alias{EMstep_PADE}
\title{EM using Matlab algorithm for matrix exponential in combination with Armadillo}
\usage{
EMstep_PADE(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{nuisance parameter}

\item{alpha}{initial probalities}

\item{S}{sub-intensity}

\item{obs}{the observations}

\item{weight}{the weights for the observations}

\item{rcens}{censored observations}

\item{rcweight}{the weights for the censored observations}
}
\description{
EM using Matlab algorithm for matrix exponential in combination with Armadillo
}
