% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_RK}
\alias{EMstep_RK}
\title{EM using Runge Kutta}
\usage{
EMstep_RK(h, pi, T, obs, weight, rcens, rcweight)
}
\description{
Warning - I am assuming that pi is a matrix I would need to make sure that this is the case in the iteration or change how I iterate with it
}
