% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_RK}
\alias{EMstep_RK}
\title{EM step using Runge Kutta}
\usage{
EMstep_RK(h, pi, T, obs, weight, rcens, rcweight)
}
\description{
Computes one step of the EM algorithm by using a Runge-Kutta method of 4th order
}
