% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_RK}
\alias{EMstep_RK}
\title{EM step for phase-type using Runge-Kutta}
\usage{
EMstep_RK(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Step-length.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{obs}{The observations.}

\item{weight}{The weights for the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{The weights for the censored observations.}
}
\description{
Computes one step of the EM algorithm by using a Runge-Kutta method of fourth order.
}
