% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_RK}
\alias{EMstep_RK}
\title{EM step using Runge Kutta}
\usage{
EMstep_RK(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{step-length}

\item{alpha}{initial probalities}

\item{S}{sub-intensity}

\item{obs}{the observations}

\item{weight}{the weights for the observations}

\item{rcens}{censored observations}

\item{rcweight}{the weights for the censored observations}
}
\description{
Computes one step of the EM algorithm by using a Runge-Kutta method of 4th order
}
