% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_UNI}
\alias{EMstep_UNI}
\title{EM using Uniformization for matrix exponential}
\usage{
EMstep_UNI(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{positive parameter}

\item{alpha}{initial probalities}

\item{S}{sub-intensity}

\item{obs}{the observations}

\item{weight}{the weights for the observations}

\item{rcens}{censored observations}

\item{rcweight}{the weights for the censored observations}
}
\description{
EM using Uniformization for matrix exponential
}
