% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EMstep_UNI}
\alias{EMstep_UNI}
\title{EM for phase-type using uniformization for matrix exponential}
\usage{
EMstep_UNI(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Positive parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{obs}{The observations.}

\item{weight}{The weights for the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{The weights for the censored observations.}
}
\description{
EM for phase-type using uniformization for matrix exponential
}
