% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.sph.R
\name{evaluate,sph-method}
\alias{evaluate,sph-method}
\title{Evaluation method for sph Class}
\usage{
\S4method{evaluate}{sph}(x, subject)
}
\arguments{
\item{x}{An object of class \linkS4class{sph}.}

\item{subject}{Covariates of a single subject.}
}
\value{
A \linkS4class{ph} model.
}
\description{
Evaluation method for sph Class
}
