% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expm_terms}
\alias{expm_terms}
\title{expm terms of phase-type likelihood using uniformization}
\usage{
expm_terms(h, S, obs)
}
\arguments{
\item{h}{Positive parameter.}

\item{S}{Sub-intensity matrix.}

\item{obs}{The observations.}
}
\description{
expm terms of phase-type likelihood using uniformization
}
