% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.sph.R
\name{Fisher,sph-method}
\alias{Fisher,sph-method}
\title{Fisher Information Method for sph Class}
\usage{
\S4method{Fisher}{sph}(x, y, X, w = numeric(0))
}
\arguments{
\item{x}{An object of class \linkS4class{sph}.}

\item{y}{Independent variate.}

\item{X}{Matrix of covariates.}

\item{w}{Weights.}
}
\value{
A matrix.
}
\description{
Fisher Information Method for sph Class
}
