% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{Fisher}
\alias{Fisher}
\title{New Generic for obtaining the Fisher Information of Survival Matrix Distributions}
\usage{
Fisher(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{further parameters to be passed on}
}
\description{
Methods are available for objects of class \linkS4class{sph}
}
