% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{fit,mph-method}
\alias{fit,mph-method}
\title{Fit Method for mph Class}
\usage{
\S4method{fit}{mph}(x, y, stepsEM = 1000, equal_marginals = FALSE)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{y}{Matrix of data.}

\item{stepsEM}{Number of EM steps to be performed.}

\item{equal_marginals}{Logical. If TRUE, all marginals are fitted to be equal.}
}
\description{
Fit Method for mph Class
}
