% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{fit,ph-method}
\alias{fit,ph-method}
\title{Fit Method for ph Class}
\usage{
\S4method{fit}{ph}(x, y)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{vector or data.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Fit Method for ph Class
}
\examples{

}
