% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\docType{methods}
\name{fit,ph-method}
\alias{fit,ph-method}
\title{Fit Method for ph Class}
\usage{
\S4method{fit}{ph}(x, y, xweight = numeric(0), rcen = numeric(0),
  rcenweight = numeric(0), stepsEM = 1000)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{vector or data.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Fit Method for ph Class
}
\examples{

}
