% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\docType{class}
\name{iph-class}
\alias{iph-class}
\title{Inhomogeneous Phase Type distributions}
\value{
Class object.
}
\description{
Class of objects for inhomogeneous phase-type distributions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the phase-type distribution.}

\item{\code{gfun}}{A list comprising of the parameters.}

\item{\code{scale}}{Scale.}
}}

