% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4b.bivph.R
\name{laplace,bivph-method}
\alias{laplace,bivph-method}
\title{Laplace Method for bivph class}
\usage{
\S4method{laplace}{bivph}(x, r)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{r}{A matrix of real values.}
}
\value{
A vector containing the corresponding Laplace transform evaluations.
}
\description{
Laplace Method for bivph class
}
\examples{
obj <- bivph(dimensions = c(3, 3))
laplace(obj, matrix(c(0.5, 1), ncol = 2))
}
