% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{laplace,ph-method}
\alias{laplace,ph-method}
\title{Laplace Method for phase-type distributions}
\usage{
\S4method{laplace}{ph}(x, r)
}
\arguments{
\item{x}{An object of class \linkS4class{ph}.}

\item{r}{A vector of real values.}
}
\value{
The Laplace transform of the \linkS4class{ph}
 (or undelying \linkS4class{ph}) object at the given locations.
}
\description{
Laplace Method for phase-type distributions
}
\examples{
set.seed(123)
obj <- ph(structure = "general", dimension = 3)
laplace(obj, 3)
}
