% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linear_combination}
\alias{linear_combination}
\title{Computes PH parameters of a linear combination of vector from MPHstar}
\usage{
linear_combination(w, alpha, S, R)
}
\arguments{
\item{w}{Vector with weights.}

\item{alpha}{Initial distribution vector.}

\item{S}{Sub-intensity matrix.}

\item{R}{Reward matrix.}
}
\value{
A list of PH parameters.
}
\description{
Computes PH parameters of a linear combination of vector from MPHstar
}
