% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{logLik,ph-method}
\alias{logLik,ph-method}
\title{logLik Method for ph Class}
\usage{
\S4method{logLik}{ph}(object)
}
\arguments{
\item{object}{an object of class \linkS4class{ph}.}
}
\value{
an object of class logLik
}
\description{
logLik Method for ph Class
}
