% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{logLik,ph-method}
\alias{logLik,ph-method}
\title{logLik Method for ph Class}
\usage{
\S4method{logLik}{ph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{ph}.}
}
\value{
An object of class logLik.
}
\description{
logLik Method for ph Class
}
\examples{
obj <- iph(ph(structure = "general", dimension = 2), gfun = "weibull", gfun_pars = 2)
data <- sim(obj, n = 100)
fitted_ph <- fit(obj, data, stepsEM = 10)
logLik(fitted_ph)
}
