% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodDPH}
\alias{logLikelihoodDPH}
\title{Loglikelihood for discrete phase-type}
\usage{
logLikelihoodDPH(alpha, S, obs, weight)
}
\arguments{
\item{alpha}{Initial probabilities.}

\item{S}{Sub-transition matrix.}

\item{obs}{The observations.}

\item{weight}{The weights of the observations.}
}
\description{
Loglikelihood for discrete phase-type
}
