% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMgev_RK}
\alias{logLikelihoodMgev_RK}
\title{Loglikelihood of matrix-GEV using Runge-Kutta}
\usage{
logLikelihoodMgev_RK(h, alpha, S, beta, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Step-length.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity.}

\item{beta}{Parameter of transformation}

\item{obs}{The observations.}

\item{weight}{Weight of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{Weight of the censored observations.}
}
\description{
Loglikelihood for a sample.
}
