% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMGEVD_RK}
\alias{logLikelihoodMGEVD_RK}
\title{Loglikelihood of matrix GEV using RK}
\usage{
logLikelihoodMGEVD_RK(h, pi, T, beta, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{step-lenght}

\item{pi}{initial probabilities}

\item{T}{sub-intensity}

\item{beta}{parametor of transformation}

\item{obs}{the observations}

\item{weight}{weight of the observations}

\item{rcens}{censored observations}

\item{rcweight}{weight of the censored observations}
}
\description{
Loglikelihood for a sample
}
