% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMgompertz_PADE}
\alias{logLikelihoodMgompertz_PADE}
\title{Loglikelihood of matrix-Gompertz using Pade}
\usage{
logLikelihoodMgompertz_PADE(h, alpha, S, beta, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Nuisance parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{beta}{Inhomogeneity parameter.}

\item{obs}{The observations.}

\item{weight}{The weights of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{The weights of the censored observations.}
}
\description{
Loglikelihood for a sample.
}
