% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMgompertz_RK}
\alias{logLikelihoodMgompertz_RK}
\title{Loglikelihood of matrix Gompertz using RK}
\usage{
logLikelihoodMgompertz_RK(h, alpha, S, beta, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{step-length}

\item{alpha}{initial probabilities}

\item{S}{sub-intensity}

\item{beta}{parameter of transformation}

\item{obs}{the observations}

\item{weight}{weight of the observations}

\item{rcens}{censored observations}

\item{rcweight}{weight of the censored observations}
}
\description{
Loglikelihood for a sample
}
