% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMLogNormal_RK}
\alias{logLikelihoodMLogNormal_RK}
\title{Loglikelihood of matrix LogNormal using RK}
\usage{
logLikelihoodMLogNormal_RK(h, pi, T, beta, obs, weight, rcens, rcweight)
}
\description{
Loglikelihood of matrix LogNormal using RK
}
