% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMlognormal_RK}
\alias{logLikelihoodMlognormal_RK}
\title{Loglikelihood of matrix-lognormal using Runge-Kutta}
\usage{
logLikelihoodMlognormal_RK(h, alpha, S, beta, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Step-length.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{beta}{Parameter of transformation.}

\item{obs}{The observations.}

\item{weight}{Weights of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{Weights of the censored observations.}
}
\description{
Loglikelihood for a sample.
}
