% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodMweibull_PADE}
\alias{logLikelihoodMweibull_PADE}
\title{Loglikelihood of matrix-Weibull using Pade}
\usage{
logLikelihoodMweibull_PADE(h, alpha, S, beta, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{nuisance parameter}

\item{alpha}{initial probabilities}

\item{S}{sub-intensity}

\item{beta}{in-homogeneity parameter}

\item{obs}{the observations}

\item{weight}{weight of the observations}

\item{rcens}{censored observations}

\item{rcweight}{weight of the censored observations}
}
\description{
Loglikelihood for a sample
}
