% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_MoE}
\alias{logLikelihoodPH_MoE}
\title{Loglikelihood for PH-MoE}
\usage{
logLikelihoodPH_MoE(alpha1, alpha2, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{alpha1}{Initial probabilities for non-censored data.}

\item{alpha2}{Initial probabilities for censored data.}

\item{S}{Sub-intensity matrix.}

\item{obs}{The observations.}

\item{weight}{The weights of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{The weights of the censored observations.}
}
\description{
Loglikelihood for PH-MoE
}
