% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_RK}
\alias{logLikelihoodPH_RK}
\title{Loglikelihood using RK}
\usage{
logLikelihoodPH_RK(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{step-length}

\item{alpha}{initial probabilities}

\item{S}{sub-intensity}

\item{obs}{the observations}

\item{weight}{weight of the observations}

\item{rcens}{censored observations}

\item{rcweight}{weight of the censored observations}
}
\description{
Loglikelihood for a sample
}
