% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_RK}
\alias{logLikelihoodPH_RK}
\title{Loglikelihood of phase-type using Runge-Kutta}
\usage{
logLikelihoodPH_RK(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Step-length.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{obs}{The observations.}

\item{weight}{Weights of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{Weights of the censored observations.}
}
\description{
Loglikelihood for a sample.
}
