% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_RKs}
\alias{logLikelihoodPH_RKs}
\title{Loglikelihood of PI with phase-type using Runge-Kutta}
\usage{
logLikelihoodPH_RKs(h, alpha, S, obs, weight, rcens, rcweight, scale1, scale2)
}
\arguments{
\item{h}{Step-length.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{obs}{The observations.}

\item{weight}{Weights of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{Weights of the censored observations.}

\item{scale1}{Scale for observations.}

\item{scale2}{Scale for censored observations.}
}
\description{
Loglikelihood for a sample.
}
