% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_RKs}
\alias{logLikelihoodPH_RKs}
\title{Loglikelihood of PH using RK}
\usage{
logLikelihoodPH_RKs(h, alpha, S, obs, weight, rcens, rcweight, scale1, scale2)
}
\arguments{
\item{h}{positive parameter}

\item{alpha}{initial probabilities}

\item{S}{sub-intensity}

\item{obs}{the observations}

\item{weight}{weight of the observations}

\item{rcens}{censored observations}

\item{rcweight}{weight of the censored observations}

\item{scale1}{scale for observations}

\item{scale2}{scale for censored observations}
}
\description{
Loglikelihood for a sample
}
