% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLikelihoodPH_UNI}
\alias{logLikelihoodPH_UNI}
\title{Loglikelihood of phase-type using uniformization}
\usage{
logLikelihoodPH_UNI(h, alpha, S, obs, weight, rcens, rcweight)
}
\arguments{
\item{h}{Positive parameter.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity.}

\item{obs}{The observations.}

\item{weight}{Weight of the observations.}

\item{rcens}{Censored observations.}

\item{rcweight}{Weight of the censored observations.}
}
\description{
Loglikelihood for a sample.
}
