% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{LRT}
\alias{LRT}
\title{New Generic for doing a likelihood ratio test between two Matrix Distribution models}
\usage{
LRT(x, y, ...)
}
\arguments{
\item{x, y}{objects of the model class.}

\item{...}{further parameters to be passed on}
}
\value{
a likelihood ratio test result.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
