% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{LRT}
\alias{LRT}
\title{New generic for likelihood ratio test between two matrix distribution models}
\usage{
LRT(x, y, ...)
}
\arguments{
\item{x, y}{Objects of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
A likelihood ratio test result.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
