% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{m_exp_sum}
\alias{m_exp_sum}
\title{Computes exp(Sx) via series representation}
\usage{
m_exp_sum(x, n, pow_vector, a)
}
\arguments{
\item{x}{A number.}

\item{n}{An integer.}

\item{pow_vector}{A vector.}

\item{a}{A number.}
}
\description{
Computes exp(Sx) via series representation
}
