% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_product}
\alias{matrix_product}
\title{Product of two matrices}
\usage{
matrix_product(a, b)
}
\arguments{
\item{a}{matrix}

\item{b}{matrix}
}
\value{
Computes C = A * B
}
\description{
Product of two matrices
}
