\name{matrixdist-package}
\alias{matrixdist-package}
\alias{matrixdist}
\docType{package}
\title{
  Statistics for Matrix Distributions
}
\description{
  This package is concerned with homogeneous and inhomogeneous phase--type distributions. 
 Methods for functional evaluation, simulation and estimation using the EM algorithm are provided.
}
\author{
Martin Bladt and Jorge Yslas.

Maintainer: Martin Bladt <martinbladt@gmail.com>
}
\references{
  Asmussen, S., Nerman, O., & Olsson, M. (1996). Fitting phase-type distributions via the EM
  algorithm. Scandinavian Journal of Statistics, 419-441.
  
  Olsson, M. (1996). Estimation of phase-type distributions from censored data. Scandinavian journal of statistics, 443-460.
  
  Albrecher, H., & Bladt, M. (2019). Inhomogeneous phase-type distributions and heavy tails. Journal of Applied Probability, 56(4), 1044-1064.
  
  Albrecher, H., Bladt, M., & Yslas, J. (2020). Fitting inhomogeneous Phase-Type distributions to data: The univariate and the multivariate case. Scandinavian Journal of Statistics.
  
  Bladt, M., & Yslas, J. (2020). Inhomogeneous Markov Survival Regression Models. arXiv:2011.03219.
}
\keyword{ matrixdist }
