% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{merge_matrices}
\alias{merge_matrices}
\title{Merges the matrices S11, S12 and S22 into a sub-intensity matrix}
\usage{
merge_matrices(S11, S12, S22)
}
\arguments{
\item{S11}{A sub-intensity matrix.}

\item{S12}{A matrix.}

\item{S22}{A sub-intensity matrix.}
}
\value{
A sub-intensity matrix.
}
\description{
Merges the matrices S11, S12 and S22 into a sub-intensity matrix
}
