% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgevcdf}
\alias{mgevcdf}
\title{Matrix GEV cdf}
\usage{
mgevcdf(x, alpha, S, mu, sigma, xi, lower_tail = TRUE)
}
\arguments{
\item{x}{non-negative value}

\item{alpha}{Initial probabilities}

\item{S}{sub-intensity matrix}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{xi}{shape parameter}

\item{lower_tail}{cdf or tail}
}
\value{
The cdf (tail) at \code{x}
}
\description{
Computes the cdf (tail) of a matrix GEV distribution with parameters \code{alpha}, \code{S} and \code{beta} at \code{x}
}
