% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mGEVDcdf}
\alias{mGEVDcdf}
\title{Matrix GEV cdf}
\usage{
mGEVDcdf(x, pi, T, mu, sigma, xi, lower_tail = TRUE)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{shape parameter}
}
\value{
The cdf (tail) at \code{x}
}
\description{
Computes the cdf (tail) of a matrix GEV distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
}
