% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mGEVDden}
\alias{mGEVDden}
\title{Matrix GEV density}
\usage{
mGEVDden(x, pi, T, mu, sigma, xi)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{mu}{location parameter}

\item{sigma}{scale parameter}

\item{xi}{shape parameter}
}
\value{
The density at \code{x}
}
\description{
Computes the density of a matrix GEV distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
Dont allow for atoms in zero
}
