% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgevden}
\alias{mgevden}
\title{Matrix GEV density}
\usage{
mgevden(x, alpha, S, mu, sigma, xi)
}
\arguments{
\item{x}{Non-negative value.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{mu}{Location parameter.}

\item{sigma}{Scale parameter.}

\item{xi}{Shape parameter.}
}
\value{
The density at \code{x}.
}
\description{
Computes the density of a matrix GEV distribution with parameters
 \code{alpha}, \code{S} and \code{beta} at \code{x}.
Dont allow for atoms in zero
}
