% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgevden}
\alias{mgevden}
\title{Matrix-GEV density}
\usage{
mgevden(x, alpha, S, beta)
}
\arguments{
\item{x}{Non-negative value.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{beta}{Transformation parameters.}
}
\value{
The density at \code{x}.
}
\description{
Computes the density of a matrix-GEV distribution with parameters
\code{alpha}, \code{S} and \code{beta} at \code{x}.
Does not allow for atoms in zero.
}
