% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgompertzden}
\alias{mgompertzden}
\title{Matrix Gompertz density}
\usage{
mgompertzden(x, pi, T, beta)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{parameter}
}
\value{
The density at \code{x}
}
\description{
Computes the density of a matrix Gompertz distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
}
