% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{minimum,iph,iph-method}
\alias{minimum,iph,iph-method}
\title{Minimum method for inhomogeneous phase-type distributions}
\usage{
\S4method{minimum}{iph,iph}(x1, x2)
}
\arguments{
\item{x1}{An object of class \linkS4class{iph}.}

\item{x2}{An object of class \linkS4class{iph}.}
}
\value{
An object of class \linkS4class{iph}.
}
\description{
Minimum method for inhomogeneous phase-type distributions
}
\examples{
iph1 <- iph(ph(structure = "general", dimension = 3), gfun = "weibull", gfun_pars = 2)
iph2 <- iph(ph(structure = "gcoxian", dimension = 5), gfun = "weibull", gfun_pars = 2)
iph_min <- minimum(iph1, iph2)
iph_min
}
