% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{minimum}
\alias{minimum}
\title{New Generic for Minimum of two Matrix Distributions}
\usage{
minimum(x1, x2, ...)
}
\arguments{
\item{x1}{an object of the model class.}

\item{x2}{an object of the model class.}

\item{...}{further parameters to be passed on}
}
\value{
A realization from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
