% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{minimum}
\alias{minimum}
\title{New generic for minimum of two matrix distributions}
\usage{
minimum(x1, x2, ...)
}
\arguments{
\item{x1}{An object of the model class.}

\item{x2}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
An object of the model class.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
