% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mloglogisticden}
\alias{mloglogisticden}
\title{Matrix Log-logistic density}
\usage{
mloglogisticden(x, alpha, S, beta)
}
\arguments{
\item{x}{Non-negative value.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{beta}{Parameters.}
}
\value{
The density at \code{x}.
}
\description{
Computes the density of a matrix Log-Logistic distribution with parameters
 \code{alpha}, \code{S} and \code{beta} at \code{x}.
}
