% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{MoE,dph-method}
\alias{MoE,dph-method}
\title{MoE method for dph Class}
\usage{
\S4method{MoE}{dph}(
  x,
  formula,
  data,
  alpha_vecs = NULL,
  weight = numeric(0),
  stepsEM = 1000,
  every = 10,
  rand_init = TRUE,
  maxWts = 1000
)
}
\arguments{
\item{x}{An object of class \linkS4class{dph}.}

\item{formula}{A regression formula.}

\item{data}{A data frame.}

\item{alpha_vecs}{Matrix of initial probabilities.}

\item{weight}{Vector of weights.}

\item{stepsEM}{Number of EM steps to be performed.}

\item{every}{Number of iterations between likelihood display updates.}

\item{rand_init}{Random initiation in the R-step.}

\item{maxWts}{Maximal number of weights in the nnet function.}
}
\value{
An object of class \linkS4class{sph}.
}
\description{
MoE method for dph Class
}
