% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.mdph.R
\name{MoE,mdph-method}
\alias{MoE,mdph-method}
\title{MoE Method for mdph Class}
\usage{
\S4method{MoE}{mdph}(
  x,
  formula,
  y,
  data,
  alpha_vecs = NULL,
  weight = numeric(0),
  stepsEM = 1000,
  every = 10,
  rand_init = TRUE
)
}
\arguments{
\item{x}{An object of class \linkS4class{mdph}.}

\item{formula}{A regression formula.}

\item{y}{A matrix of observations.}

\item{data}{A data frame of covariates.}

\item{alpha_vecs}{Matrix of initial probabilities.}

\item{weight}{Vector of weights.}

\item{stepsEM}{Number of EM steps to be performed.}

\item{every}{Number of iterations between likelihood display updates.}

\item{rand_init}{Random initiation in the R-step.}
}
\value{
An object of class \linkS4class{sph}.
}
\description{
MoE Method for mdph Class
}
