% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{moment,ph-method}
\alias{moment,ph-method}
\title{Moment method for phase-type distributions}
\usage{
\S4method{moment}{ph}(x, k = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{ph}.}

\item{k}{A positive integer (moment order).}
}
\value{
The raw moment of the \linkS4class{ph} (or underlying \linkS4class{ph}) object.
}
\description{
Moment method for phase-type distributions
}
\examples{
set.seed(123)
obj <- ph(structure = "general", dimension = 3)
moment(obj, 2)
}
