% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{moment}
\alias{moment}
\title{New Generic for Moment of Matrix Distributions}
\usage{
moment(x, ...)
}
\arguments{
\item{x}{an object of the model class.}

\item{...}{further parameters to be passed on}
}
\value{
A realization from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
