% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mParetocdf}
\alias{mParetocdf}
\title{Matrix Pareto cdf}
\usage{
mParetocdf(x, pi, T, beta, lower_tail = TRUE)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{shape parameter}
}
\value{
The cdf (tail) at \code{x}
}
\description{
Computes the cdf (tail) of a matrix Pareto distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
}
