% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mparetoden}
\alias{mparetoden}
\title{Matrix Pareto density}
\usage{
mparetoden(x, pi, T, beta)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{scale parameter}
}
\value{
The density at \code{x}
}
\description{
Computes the density of a matrix Pareto distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
}
