% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mparetoden}
\alias{mparetoden}
\title{Matrix-Pareto density}
\usage{
mparetoden(x, alpha, S, beta)
}
\arguments{
\item{x}{Non-negative value.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{beta}{Scale parameter.}
}
\value{
The density at \code{x}.
}
\description{
Computes the density of a matrix-Pareto distribution with parameters
\code{alpha}, \code{S} and \code{beta} at \code{x}.
}
