% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\docType{class}
\name{mph-class}
\alias{mph-class}
\title{Multivariate phase-type distributions}
\value{
Class object.
}
\description{
Class of objects for multivariate phase-type distributions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the phase type distribution.}

\item{\code{pars}}{A list comprising of the parameters.}

\item{\code{fit}}{A list containing estimation information.}
}}

