% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mWeibullcdf}
\alias{mWeibullcdf}
\title{Matrix Weibull cdf}
\usage{
mWeibullcdf(x, pi, T, beta, lower_tail)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{shape parameter}
}
\value{
The cdf (tail) at \code{x}
}
\description{
Computes the cdf (tail) of a matrix Weibull distribution with parameters \code{pi}, \code{T} and \code{beta} at \code{x}
}
