% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mweibullcdf}
\alias{mweibullcdf}
\title{Matrix Weibull cdf}
\usage{
mweibullcdf(x, alpha, S, beta, lower_tail)
}
\arguments{
\item{x}{non-negative value}

\item{alpha}{Initial probabilities}

\item{S}{sub-intensity matrix}

\item{beta}{shape parameter}

\item{lower_tail}{cdf or tail}
}
\value{
The cdf (tail) at \code{x}
}
\description{
Computes the cdf (tail) of a matrix Weibull distribution with parameters \code{alpha}, \code{S} and \code{beta} at \code{x}
}
