% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{pgf}
\alias{pgf}
\title{New Generic for pgf of Matrix Distributions}
\usage{
pgf(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
Pgf transform of the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{dph}.
}
