% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\docType{class}
\name{ph-class}
\alias{ph-class}
\title{Phase Type distributions}
\value{
Class object
}
\description{
Class of objects for phase type distributions
}
\section{Slots}{

\describe{
\item{\code{name}}{name of the phase type distribution.}

\item{\code{pars}}{a list comprising of the parameters.}

\item{\code{fit}}{a list containing estimation information.}
}}

