% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{ph}
\alias{ph}
\title{Constructor Function for phase type distributions}
\usage{
ph(alpha = NA, S = NA, structure = NA, order = 3)
}
\arguments{
\item{alpha}{a probability vector.}

\item{S}{an intensity matrix.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Constructor Function for phase type distributions
}
