% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph.R
\name{ph}
\alias{ph}
\title{Constructor Function for phase type distributions}
\usage{
ph(pi = c(1), T = as.matrix(-1))
}
\arguments{
\item{pi}{a probability vector.}

\item{T}{an intensity matrix.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Constructor Function for phase type distributions
}
