% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{ph}
\alias{ph}
\title{Constructor function for phase-type distributions}
\usage{
ph(alpha = NULL, S = NULL, structure = NULL, dimension = 3)
}
\arguments{
\item{alpha}{A probability vector.}

\item{S}{A sub-intensity matrix.}

\item{structure}{A valid ph structure: \code{"general"}, \code{"coxian"},
\code{"hyperexponential"}, \code{"gcoxian"}, or \code{"gerlang"}.}

\item{dimension}{The dimension of the ph structure (if structure is provided).}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Constructor function for phase-type distributions
}
\examples{
ph(structure = "gcoxian", dimension = 5)
ph(alpha = c(.5, .5), S = matrix(c(-1, .5, .5, -1), 2, 2))
}
