% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phcdf}
\alias{phcdf}
\title{Phase-type cdf or tail}
\usage{
phcdf(x, pi, T, lower_tail = TRUE)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}
}
\value{
The cdf (tail) at \code{x}
}
\description{
Computes the cdf of phase-type distribution with parameters \code{pi} and \code{T} at \code{x}
}
