% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phcdf}
\alias{phcdf}
\title{Phase-type cdf or tail}
\usage{
phcdf(x, alpha, S, lower_tail = TRUE)
}
\arguments{
\item{x}{non-negative value}

\item{alpha}{Initial probabilities}

\item{S}{sub-intensity matrix}

\item{lower_tail}{cdf or tail}
}
\value{
The cdf (tail) at \code{x}
}
\description{
Computes the cdf of phase-type distribution with parameters \code{alpha} and \code{S} at \code{x}
}
