% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{phdensity}
\alias{phdensity}
\title{Phase-type density}
\usage{
phdensity(x, pi, T)
}
\arguments{
\item{x}{non-negative value}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}
}
\value{
The density at \code{x}
}
\description{
Computes the density of phase-type distribution with parameters \code{pi} and \code{T} at \code{x}
}
