% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{+,ph,ph-method}
\alias{+,ph,ph-method}
\title{Sum Method for phase-type distributions}
\usage{
\S4method{+}{ph,ph}(e1, e2)
}
\arguments{
\item{e1}{An object of class \linkS4class{ph}.}

\item{e2}{An object of class \linkS4class{ph}.}
}
\value{
An object of class \linkS4class{ph}.
}
\description{
Sum Method for phase-type distributions
}
\examples{
ph1 <- ph(structure = "general", dimension = 3)
ph2 <- ph(structure = "gcoxian", dimension = 5)
ph_sum <- ph1 + ph2
ph_sum
}
