% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.ph.R
\name{quan,ph-method}
\alias{quan,ph-method}
\title{Quantile Method for phase type distributions}
\usage{
\S4method{quan}{ph}(x, p = seq(0, 1, length.out = 10))
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{p}{a vector of probabilities.}
}
\value{
A list containing the probabilities and corresponding quantile evaluations.
}
\description{
Quantile Method for phase type distributions
}
\examples{
obj <- ph(structure = "general")
quan(obj, c(0.5, 0.9, 0.99))
}
