% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b.iph_regression.R
\name{reg,ph-method}
\alias{reg,ph-method}
\title{Regression method for ph Class}
\usage{
\S4method{reg}{ph}(
  x,
  y,
  weight = numeric(0),
  rcen = numeric(0),
  rcenweight = numeric(0),
  X = numeric(0),
  B0 = numeric(0),
  stepsEM = 1000,
  methods = c("RK", "UNI"),
  rkstep = NA,
  uni_epsilon = NA,
  optim_method = "BFGS",
  maxit = 50,
  reltol = 1e-08,
  every = 10
)
}
\arguments{
\item{x}{An object of class \linkS4class{ph}.}

\item{y}{Vector or data.}

\item{weight}{Vector of weights.}

\item{rcen}{Vector of right-censored observations.}

\item{rcenweight}{Vector of weights for right-censored observations.}

\item{X}{Model matrix (no intercept needed).}

\item{B0}{Initial regression coefficients (optional).}

\item{stepsEM}{Number of EM steps to be performed.}

\item{methods}{Methods to use for matrix exponential calculation: \code{RM}, \code{UNI}, or \code{PADE}.}

\item{rkstep}{Runge-Kutta step size (optional).}

\item{uni_epsilon}{Epsilon parameter for uniformization method.}

\item{optim_method}{Method to use in gradient optimization.}

\item{maxit}{Maximum number of iterations when optimizing g function.}

\item{reltol}{Relative tolerance when optimizing g function.}

\item{every}{Number of iterations between likelihood display updates.}
}
\value{
An object of class \linkS4class{sph}.
}
\description{
Regression method for ph Class
}
\examples{
set.seed(1)
obj <- iph(ph(structure = "general", dimension = 2), gfun = "weibull", gfun_pars = 2)
data <- sim(obj, n = 100)
X <- runif(100)
reg(x = obj, y = data, X = X, stepsEM = 10)
}
