% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2b.iph_regression.R
\name{reg,ph-method}
\alias{reg,ph-method}
\title{Regression Method for ph Class}
\usage{
\S4method{reg}{ph}(
  x,
  y,
  weight = numeric(0),
  rcen = numeric(0),
  rcenweight = numeric(0),
  X = numeric(0),
  B0 = numeric(0),
  stepsEM = 1000,
  methods = c("RK", "UNI"),
  rkstep = NA,
  uni_epsilon = NA,
  optim_method = "BFGS",
  maxit = 50,
  reltol = 1e-08,
  every = 10
)
}
\arguments{
\item{x}{an object of class \linkS4class{ph}.}

\item{y}{vector or data.}

\item{weight}{vector of weights.}

\item{rcen}{vector of right-censored observations}

\item{rcenweight}{vector of weights for right-censored observations.}

\item{X}{model matrix (no intercept needed).}

\item{B0}{initial regression coefficients (optional).}

\item{stepsEM}{number of EM steps to be performed.}

\item{methods}{methods to use for matrix exponential calculation: RM, UNI or PADE}

\item{rkstep}{Runge-Kutta step size (optional)}

\item{uni_epsilon}{epsilon parameter for uniformization method}

\item{optim_method}{method to use in gradient optimization}

\item{maxit}{maximum number of iterations when optimizing g function.}

\item{reltol}{relative tolerance when optimizing g function.}

\item{every}{number of iterations between likelihood display updates.}
}
\value{
An object of class \linkS4class{sph}.
}
\description{
Regression Method for ph Class
}
