% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{reg}
\alias{reg}
\title{New Generic for Regression with Matrix Distributions}
\usage{
reg(x, y, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{y}{A vector of data.}

\item{...}{Further parameters to be passed on.}
}
\value{
An object of the fitted model class.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
