% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reversTransformData}
\alias{reversTransformData}
\title{Applies the inverse of the GEV but giving back the vector in reverse order}
\usage{
reversTransformData(observations, weights, beta)
}
\arguments{
\item{observations}{The observations.}

\item{weights}{Weights of the observations.}

\item{beta}{Parameters of the GEV.}
}
\description{
Used for EM step.
}
