% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{riph}
\alias{riph}
\title{Random inhomogeneous phase-type}
\usage{
riph(n, dist_type, pi, T, beta)
}
\arguments{
\item{n}{Sample size}

\item{dist_type}{Type of IPH: "Pareto", "Weibull", "Gompertz"}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{beta}{Parameter of the transformation}
}
\value{
The simulated sample
}
\description{
Generates a sample of size \code{n} from an inhomogeneous phase-type distribution with parameters \code{pi}, \code{T} and \code{beta}
}
