% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{riph}
\alias{riph}
\title{Random inhomogeneous phase-type}
\usage{
riph(n, dist_type, alpha, S, beta)
}
\arguments{
\item{n}{Sample size.}

\item{dist_type}{Type of IPH.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{beta}{Parameter of the transformation.}
}
\value{
The simulated sample.
}
\description{
Generates a sample of size \code{n} from an inhomogeneous phase-type 
distribution with parameters \code{alpha}, \code{S} and \code{beta}.
}
