% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmatrixGEVD}
\alias{rmatrixGEVD}
\title{Random matrix GEVD}
\usage{
rmatrixGEVD(n, pi, T, mu, sigma, xi = 0)
}
\arguments{
\item{n}{Sample size}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}

\item{mu}{Location parameter}

\item{sigma}{Scale parameter}

\item{xi}{Shape parameter: Default 0 which corresponds to the Gumbel case}
}
\value{
The simulated sample
}
\description{
Generates a sample of size \code{n} from an inhomogeneous phase-type distribution with parameters \code{pi}, \code{T} and \code{beta}
}
