% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rMPHstar}
\alias{rMPHstar}
\title{Simulate a MPH* random vector}
\usage{
rMPHstar(n, alpha, S, R)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Initial probabilities.}

\item{S}{Sub-intensity matrix.}

\item{R}{Reward matrix.}
}
\value{
The simulated sample.
}
\description{
Generates a sample of size \code{n} from a MPH* distribution with parameters
 \code{alpha}, \code{S} and \code{R}.
}
