% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rphasetype}
\alias{rphasetype}
\title{Random phase-type}
\usage{
rphasetype(n, pi, T)
}
\arguments{
\item{n}{Sample size}

\item{pi}{Initial probabilities}

\item{T}{sub-intensity matrix}
}
\value{
The simulated sample
}
\description{
Generates a sample of size \code{n} from a phase-type distribution with parameters \code{pi} and \code{T}
}
