% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rphasetype}
\alias{rphasetype}
\title{Simulate phase-type}
\usage{
rphasetype(n, alpha, S)
}
\arguments{
\item{n}{Sample size.}

\item{alpha}{Vector of initial probabilities.}

\item{S}{Sub-intensity matrix.}
}
\value{
Simulated sample.
}
\description{
Generates a sample of size \code{n} from a phase-type distribution with
parameters \code{alpha} and \code{S}.
}
