% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runge_kutta}
\alias{runge_kutta}
\title{Runge-Kutta for the calculation of the a and b vectors and the c matrix in a EM step}
\usage{
runge_kutta(avector, bvector, cmatrix, dt, h, S, s)
}
\arguments{
\item{avector}{The a vector.}

\item{bvector}{The b vector.}

\item{cmatrix}{The c matrix.}

\item{dt}{The increment.}

\item{h}{Step-length.}

\item{S}{Sub-intensity matrix.}

\item{s}{Exit rates.}
}
\description{
Performs the Runge-Kutta method of fourth order.
}
