% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
<<<<<<< HEAD
\docType{methods}
=======
>>>>>>> bfd4a4437aecc2d9b403fc560a9345bfda8bd07c
\name{show,iph-method}
\alias{show,iph-method}
\title{Show Method for inhomogeneous phase type distributions}
\usage{
\S4method{show}{iph}(object)
}
\arguments{
\item{x}{an object of class \linkS4class{iph}.}
}
\description{
Show Method for inhomogeneous phase type distributions
}
\examples{

}
