% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.iph.R
\name{show,iph-method}
\alias{show,iph-method}
\title{Show Method for inhomogeneous phase-type distributions}
\usage{
\S4method{show}{iph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{iph}.}
}
\description{
Show Method for inhomogeneous phase-type distributions
}
