% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{show,mph-method}
\alias{show,mph-method}
\title{Show Method for multivariate phase-type distributions}
\usage{
\S4method{show}{mph}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{mph}.}
}
\description{
Show Method for multivariate phase-type distributions
}
