% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.mph.R
\name{sim,mph-method}
\alias{sim,mph-method}
\title{Simulation method for multivariate phase-type distributions}
\usage{
\S4method{sim}{mph}(x, n = 1000, equal_marginals = 0)
}
\arguments{
\item{x}{An object of class \linkS4class{mph}.}

\item{n}{Length of realization.}

\item{equal_marginals}{Non-negative integer. If positive, it specifies
the number of marginals to simulate from, all from the first matrix.}
}
\value{
A realization of a multivariate phase-type distribution.
}
\description{
Simulation method for multivariate phase-type distributions
}
\examples{
obj <- mph(structure = c("general", "general"))
sim(obj, 100)
}
