% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{sim}
\alias{sim}
\title{New generic for simulating matrix distributions}
\usage{
sim(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
A realization from the matrix distribution.
}
\description{
Methods are available for objects of class \linkS4class{ph}.
}
