% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_linear_system}
\alias{solve_linear_system}
\title{Solves a system with multiple right hand sides}
\usage{
solve_linear_system(A1, B)
}
\arguments{
\item{A1}{a matrix}

\item{B}{a matrix}
}
\description{
AX=B which can be decomposed as LUX=B and finds X.
When B is the identity matrix the solution is the inverse of A
}
