% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.sph.R
\docType{class}
\name{sph-class}
\alias{sph-class}
\title{Survival Analysis for Phase Type distributions}
\value{
Class object
}
\description{
Class of objects for inhomogeneous phase-type distributions
}
\section{Slots}{

\describe{
\item{\code{coefs}}{Coefficients of the survival regression object.}

\item{\code{type}}{Type of survival object.}
}}

