% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.sph.R
\name{sph}
\alias{sph}
\title{Constructor function for survival phase-type objects}
\usage{
sph(x = NULL, coefs = list(B = numeric(0), C = numeric(0)), type = "reg")
}
\arguments{
\item{x}{An object of class \linkS4class{ph}.}

\item{coefs}{Coefficients of the survival regression object.}

\item{type}{Type of survival object.}
}
\value{
An object of class \linkS4class{sph}.
}
\description{
Constructor function for survival phase-type objects
}
