% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sumPH}
\alias{sumPH}
\title{Computes the initial distribution and sub-intensity of the sum of PH}
\usage{
sumPH(alpha1, S1, alpha2, S2)
}
\arguments{
\item{alpha1}{initial distribution}

\item{S1}{sub-intensity}

\item{alpha2}{initial distribution}

\item{S2}{sub-intensity}
}
\description{
Computes the initial distribution and sub-intensity of the sum of PH
}
