% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0.generics.R
\name{TVR}
\alias{TVR}
\title{New Generic for the transformation via rewards of a phase-type distribution}
\usage{
TVR(x, ...)
}
\arguments{
\item{x}{An object of the model class.}

\item{...}{Further parameters to be passed on.}
}
\value{
An object of the model class.
}
\description{
Methods are available for objects of class \linkS4class{ph}
}
