% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tvr_ph}
\alias{tvr_ph}
\title{Performs TVR for phase-type distributions}
\usage{
tvr_ph(alpha, S, R)
}
\arguments{
\item{alpha}{Initial distribution vector.}

\item{S}{Sub-intensity matrix.}

\item{R}{Reward vector.}
}
\value{
A list of phase-type parameters.
}
\description{
Performs TVR for phase-type distributions
}
