% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.dph.R
\name{var,dph-method}
\alias{var,dph-method}
\title{Var Method for discrete phase-type distributions}
\usage{
\S4method{var}{dph}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{dph}.}
}
\value{
The variance of the \linkS4class{dph} object.
}
\description{
Var Method for discrete phase-type distributions
}
\examples{
set.seed(123)
dph1 <- dph(structure = "general", dimension = 3)
var(dph1)
}
