% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.MPHstar.R
\name{var,MPHstar-method}
\alias{var,MPHstar-method}
\title{Var Method for MPHstar class}
\usage{
\S4method{var}{MPHstar}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{MPHstar}.}
}
\value{
The covariance matrix of the MPHstar distribution.
}
\description{
Var Method for MPHstar class
}
\examples{
obj <- MPHstar(structure = "general")
var(obj)
}
