% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_of_matrices}
\alias{vector_of_matrices}
\title{Computes the elements S^n / n! until the a given size}
\usage{
vector_of_matrices(vect, S, a, vect_size)
}
\arguments{
\item{vect}{A vector.}

\item{S}{Sub-intensity matrix.}

\item{a}{A number.}

\item{vect_size}{Size of vector.}
}
\description{
Computes the elements S^n / n! until the a given size
}
