% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vector_of_matrices_2}
\alias{vector_of_matrices_2}
\title{Computes the elements S^n / n! until given value of n}
\usage{
vector_of_matrices_2(vect, S, vect_size)
}
\arguments{
\item{vect}{A vector.}

\item{S}{Sub-intensity matrix.}

\item{vect_size}{Size of vector.}
}
\description{
Computes the elements S^n / n! until given value of n
}
